% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_indicators.R
\name{country_indicators}
\alias{country_indicators}
\title{Retrieves country indicators data from HDR Data API}
\usage{
country_indicators(country = NULL, year = NULL, indicator = NULL)
}
\arguments{
\item{country}{A character string specifying the country ISO code for which indicators data is requested.}

\item{year}{An integer specifying the year for which indicators data is requested.}
}
\value{
A data frame containing the country indicators data.
}
\description{
This function retrieves country indicators data from the HDR Data API based on the specified country and year.
}
\examples{
\dontrun{
country_indicators("BRA", 2010, "hdi")
}
}
