\name{HDR-package}
\alias{HDR-package}
\alias{HDR}
\docType{package}
\title{
Processing of hyperspectral reflectance data
}
\description{
Package contains two functions for processing hyperspectral reflectance data: 1) generates average reflection in several broadband spectral regions and 2) identifies spectral regions with low SN ratio
}
\details{
\tabular{ll}{
Package: \tab HDR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-23\cr
License: \tab What license is it under?\cr
}
Just directly call the functions on your hyperspectral data. 
}
\author{
Hand D Roelofsen, KWR Watercycle Research Insitute. Nieuwegein, the Netherlands

Maintainer: HDR <hans.roelofsen@kwrwater.nl>
}
\references{
none
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
