% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr_indicators.R
\docType{data}
\name{hdr_indicators}
\alias{hdr_indicators}
\title{HDR indicators}
\format{A data frame with variables \code{id} and \code{indicator}, the
  latter of which provides human-readable names of the indicators provided by
  the UNDP Human Development Report API.}
\usage{
hdr_indicators
}
\description{
HDR indicators
}
\keyword{datasets}

