\name{SampleHySpecRefl}
\alias{SampleHySpecRefl}
\docType{data}
\title{
Sample Hyperspectral reflectance data
}
\description{
Hyperspectral reflectance data for 100 artificial samples in 2151 spectral bands (350-2500nm)
Originally from ASD FieldSpec Pro FR field measurements by Hans D Roelofsen. Artifical noise added to obscure original data. 
}
\usage{data(SampleHySpecRefl)}
\format{
  A data frame with 100 observations on the following 2151 variables.
  \describe{
    \item{\code{X350}}{a numeric vector}
    \item{\code{X351}}{a numeric vector}
    \item{\code{X352}}{a numeric vector}
    \item{\code{X353}}{a numeric vector}
    \item{\code{X354}}{a numeric vector}
    \item{\code{X355}}{a numeric vector}
    \item{\code{X356}}{a numeric vector}
    \item{\code{X357}}{a numeric vector}
    \item{\code{X358}}{a numeric vector}
    \item{\code{X359}}{a numeric vector}
    \item{\code{X360}}{a numeric vector}
    \item{\code{X361}}{a numeric vector}
    \item{\code{X362}}{a numeric vector}
    \item{\code{X363}}{a numeric vector}
    \item{\code{X364}}{a numeric vector}
    \item{\code{X365}}{a numeric vector}
    \item{\code{X366}}{a numeric vector}
    \item{\code{X367}}{a numeric vector}
    \item{\code{X368}}{a numeric vector}
    \item{\code{X369}}{a numeric vector}
    \item{\code{X370}}{a numeric vector}
    \item{\code{X371}}{a numeric vector}
    \item{\code{X372}}{a numeric vector}
    \item{\code{X373}}{a numeric vector}
    \item{\code{X374}}{a numeric vector}
    \item{\code{X375}}{a numeric vector}
    \item{\code{X376}}{a numeric vector}
    \item{\code{X377}}{a numeric vector}
    \item{\code{X378}}{a numeric vector}
    \item{\code{X379}}{a numeric vector}
    \item{\code{X380}}{a numeric vector}
    \item{\code{X381}}{a numeric vector}
    \item{\code{X382}}{a numeric vector}
    \item{\code{X383}}{a numeric vector}
    \item{\code{X384}}{a numeric vector}
    \item{\code{X385}}{a numeric vector}
    \item{\code{X386}}{a numeric vector}
    \item{\code{X387}}{a numeric vector}
    \item{\code{X388}}{a numeric vector}
    \item{\code{X389}}{a numeric vector}
    \item{\code{X390}}{a numeric vector}
    \item{\code{X391}}{a numeric vector}
    \item{\code{X392}}{a numeric vector}
    \item{\code{X393}}{a numeric vector}
    \item{\code{X394}}{a numeric vector}
    \item{\code{X395}}{a numeric vector}
    \item{\code{X396}}{a numeric vector}
    \item{\code{X397}}{a numeric vector}
    \item{\code{X398}}{a numeric vector}
    \item{\code{X399}}{a numeric vector}
    \item{\code{X400}}{a numeric vector}
    \item{\code{X401}}{a numeric vector}
    \item{\code{X402}}{a numeric vector}
    \item{\code{X403}}{a numeric vector}
    \item{\code{X404}}{a numeric vector}
    \item{\code{X405}}{a numeric vector}
    \item{\code{X406}}{a numeric vector}
    \item{\code{X407}}{a numeric vector}
    \item{\code{X408}}{a numeric vector}
    \item{\code{X409}}{a numeric vector}
    \item{\code{X410}}{a numeric vector}
    \item{\code{X411}}{a numeric vector}
    \item{\code{X412}}{a numeric vector}
    \item{\code{X413}}{a numeric vector}
    \item{\code{X414}}{a numeric vector}
    \item{\code{X415}}{a numeric vector}
    \item{\code{X416}}{a numeric vector}
    \item{\code{X417}}{a numeric vector}
    \item{\code{X418}}{a numeric vector}
    \item{\code{X419}}{a numeric vector}
    \item{\code{X420}}{a numeric vector}
    \item{\code{X421}}{a numeric vector}
    \item{\code{X422}}{a numeric vector}
    \item{\code{X423}}{a numeric vector}
    \item{\code{X424}}{a numeric vector}
    \item{\code{X425}}{a numeric vector}
    \item{\code{X426}}{a numeric vector}
    \item{\code{X427}}{a numeric vector}
    \item{\code{X428}}{a numeric vector}
    \item{\code{X429}}{a numeric vector}
    \item{\code{X430}}{a numeric vector}
    \item{\code{X431}}{a numeric vector}
    \item{\code{X432}}{a numeric vector}
    \item{\code{X433}}{a numeric vector}
    \item{\code{X434}}{a numeric vector}
    \item{\code{X435}}{a numeric vector}
    \item{\code{X436}}{a numeric vector}
    \item{\code{X437}}{a numeric vector}
    \item{\code{X438}}{a numeric vector}
    \item{\code{X439}}{a numeric vector}
    \item{\code{X440}}{a numeric vector}
    \item{\code{X441}}{a numeric vector}
    \item{\code{X442}}{a numeric vector}
    \item{\code{X443}}{a numeric vector}
    \item{\code{X444}}{a numeric vector}
    \item{\code{X445}}{a numeric vector}
    \item{\code{X446}}{a numeric vector}
    \item{\code{X447}}{a numeric vector}
    \item{\code{X448}}{a numeric vector}
    \item{\code{X449}}{a numeric vector}
    \item{\code{X450}}{a numeric vector}
    \item{\code{X451}}{a numeric vector}
    \item{\code{X452}}{a numeric vector}
    \item{\code{X453}}{a numeric vector}
    \item{\code{X454}}{a numeric vector}
    \item{\code{X455}}{a numeric vector}
    \item{\code{X456}}{a numeric vector}
    \item{\code{X457}}{a numeric vector}
    \item{\code{X458}}{a numeric vector}
    \item{\code{X459}}{a numeric vector}
    \item{\code{X460}}{a numeric vector}
    \item{\code{X461}}{a numeric vector}
    \item{\code{X462}}{a numeric vector}
    \item{\code{X463}}{a numeric vector}
    \item{\code{X464}}{a numeric vector}
    \item{\code{X465}}{a numeric vector}
    \item{\code{X466}}{a numeric vector}
    \item{\code{X467}}{a numeric vector}
    \item{\code{X468}}{a numeric vector}
    \item{\code{X469}}{a numeric vector}
    \item{\code{X470}}{a numeric vector}
    \item{\code{X471}}{a numeric vector}
    \item{\code{X472}}{a numeric vector}
    \item{\code{X473}}{a numeric vector}
    \item{\code{X474}}{a numeric vector}
    \item{\code{X475}}{a numeric vector}
    \item{\code{X476}}{a numeric vector}
    \item{\code{X477}}{a numeric vector}
    \item{\code{X478}}{a numeric vector}
    \item{\code{X479}}{a numeric vector}
    \item{\code{X480}}{a numeric vector}
    \item{\code{X481}}{a numeric vector}
    \item{\code{X482}}{a numeric vector}
    \item{\code{X483}}{a numeric vector}
    \item{\code{X484}}{a numeric vector}
    \item{\code{X485}}{a numeric vector}
    \item{\code{X486}}{a numeric vector}
    \item{\code{X487}}{a numeric vector}
    \item{\code{X488}}{a numeric vector}
    \item{\code{X489}}{a numeric vector}
    \item{\code{X490}}{a numeric vector}
    \item{\code{X491}}{a numeric vector}
    \item{\code{X492}}{a numeric vector}
    \item{\code{X493}}{a numeric vector}
    \item{\code{X494}}{a numeric vector}
    \item{\code{X495}}{a numeric vector}
    \item{\code{X496}}{a numeric vector}
    \item{\code{X497}}{a numeric vector}
    \item{\code{X498}}{a numeric vector}
    \item{\code{X499}}{a numeric vector}
    \item{\code{X500}}{a numeric vector}
    \item{\code{X501}}{a numeric vector}
    \item{\code{X502}}{a numeric vector}
    \item{\code{X503}}{a numeric vector}
    \item{\code{X504}}{a numeric vector}
    \item{\code{X505}}{a numeric vector}
    \item{\code{X506}}{a numeric vector}
    \item{\code{X507}}{a numeric vector}
    \item{\code{X508}}{a numeric vector}
    \item{\code{X509}}{a numeric vector}
    \item{\code{X510}}{a numeric vector}
    \item{\code{X511}}{a numeric vector}
    \item{\code{X512}}{a numeric vector}
    \item{\code{X513}}{a numeric vector}
    \item{\code{X514}}{a numeric vector}
    \item{\code{X515}}{a numeric vector}
    \item{\code{X516}}{a numeric vector}
    \item{\code{X517}}{a numeric vector}
    \item{\code{X518}}{a numeric vector}
    \item{\code{X519}}{a numeric vector}
    \item{\code{X520}}{a numeric vector}
    \item{\code{X521}}{a numeric vector}
    \item{\code{X522}}{a numeric vector}
    \item{\code{X523}}{a numeric vector}
    \item{\code{X524}}{a numeric vector}
    \item{\code{X525}}{a numeric vector}
    \item{\code{X526}}{a numeric vector}
    \item{\code{X527}}{a numeric vector}
    \item{\code{X528}}{a numeric vector}
    \item{\code{X529}}{a numeric vector}
    \item{\code{X530}}{a numeric vector}
    \item{\code{X531}}{a numeric vector}
    \item{\code{X532}}{a numeric vector}
    \item{\code{X533}}{a numeric vector}
    \item{\code{X534}}{a numeric vector}
    \item{\code{X535}}{a numeric vector}
    \item{\code{X536}}{a numeric vector}
    \item{\code{X537}}{a numeric vector}
    \item{\code{X538}}{a numeric vector}
    \item{\code{X539}}{a numeric vector}
    \item{\code{X540}}{a numeric vector}
    \item{\code{X541}}{a numeric vector}
    \item{\code{X542}}{a numeric vector}
    \item{\code{X543}}{a numeric vector}
    \item{\code{X544}}{a numeric vector}
    \item{\code{X545}}{a numeric vector}
    \item{\code{X546}}{a numeric vector}
    \item{\code{X547}}{a numeric vector}
    \item{\code{X548}}{a numeric vector}
    \item{\code{X549}}{a numeric vector}
    \item{\code{X550}}{a numeric vector}
    \item{\code{X551}}{a numeric vector}
    \item{\code{X552}}{a numeric vector}
    \item{\code{X553}}{a numeric vector}
    \item{\code{X554}}{a numeric vector}
    \item{\code{X555}}{a numeric vector}
    \item{\code{X556}}{a numeric vector}
    \item{\code{X557}}{a numeric vector}
    \item{\code{X558}}{a numeric vector}
    \item{\code{X559}}{a numeric vector}
    \item{\code{X560}}{a numeric vector}
    \item{\code{X561}}{a numeric vector}
    \item{\code{X562}}{a numeric vector}
    \item{\code{X563}}{a numeric vector}
    \item{\code{X564}}{a numeric vector}
    \item{\code{X565}}{a numeric vector}
    \item{\code{X566}}{a numeric vector}
    \item{\code{X567}}{a numeric vector}
    \item{\code{X568}}{a numeric vector}
    \item{\code{X569}}{a numeric vector}
    \item{\code{X570}}{a numeric vector}
    \item{\code{X571}}{a numeric vector}
    \item{\code{X572}}{a numeric vector}
    \item{\code{X573}}{a numeric vector}
    \item{\code{X574}}{a numeric vector}
    \item{\code{X575}}{a numeric vector}
    \item{\code{X576}}{a numeric vector}
    \item{\code{X577}}{a numeric vector}
    \item{\code{X578}}{a numeric vector}
    \item{\code{X579}}{a numeric vector}
    \item{\code{X580}}{a numeric vector}
    \item{\code{X581}}{a numeric vector}
    \item{\code{X582}}{a numeric vector}
    \item{\code{X583}}{a numeric vector}
    \item{\code{X584}}{a numeric vector}
    \item{\code{X585}}{a numeric vector}
    \item{\code{X586}}{a numeric vector}
    \item{\code{X587}}{a numeric vector}
    \item{\code{X588}}{a numeric vector}
    \item{\code{X589}}{a numeric vector}
    \item{\code{X590}}{a numeric vector}
    \item{\code{X591}}{a numeric vector}
    \item{\code{X592}}{a numeric vector}
    \item{\code{X593}}{a numeric vector}
    \item{\code{X594}}{a numeric vector}
    \item{\code{X595}}{a numeric vector}
    \item{\code{X596}}{a numeric vector}
    \item{\code{X597}}{a numeric vector}
    \item{\code{X598}}{a numeric vector}
    \item{\code{X599}}{a numeric vector}
    \item{\code{X600}}{a numeric vector}
    \item{\code{X601}}{a numeric vector}
    \item{\code{X602}}{a numeric vector}
    \item{\code{X603}}{a numeric vector}
    \item{\code{X604}}{a numeric vector}
    \item{\code{X605}}{a numeric vector}
    \item{\code{X606}}{a numeric vector}
    \item{\code{X607}}{a numeric vector}
    \item{\code{X608}}{a numeric vector}
    \item{\code{X609}}{a numeric vector}
    \item{\code{X610}}{a numeric vector}
    \item{\code{X611}}{a numeric vector}
    \item{\code{X612}}{a numeric vector}
    \item{\code{X613}}{a numeric vector}
    \item{\code{X614}}{a numeric vector}
    \item{\code{X615}}{a numeric vector}
    \item{\code{X616}}{a numeric vector}
    \item{\code{X617}}{a numeric vector}
    \item{\code{X618}}{a numeric vector}
    \item{\code{X619}}{a numeric vector}
    \item{\code{X620}}{a numeric vector}
    \item{\code{X621}}{a numeric vector}
    \item{\code{X622}}{a numeric vector}
    \item{\code{X623}}{a numeric vector}
    \item{\code{X624}}{a numeric vector}
    \item{\code{X625}}{a numeric vector}
    \item{\code{X626}}{a numeric vector}
    \item{\code{X627}}{a numeric vector}
    \item{\code{X628}}{a numeric vector}
    \item{\code{X629}}{a numeric vector}
    \item{\code{X630}}{a numeric vector}
    \item{\code{X631}}{a numeric vector}
    \item{\code{X632}}{a numeric vector}
    \item{\code{X633}}{a numeric vector}
    \item{\code{X634}}{a numeric vector}
    \item{\code{X635}}{a numeric vector}
    \item{\code{X636}}{a numeric vector}
    \item{\code{X637}}{a numeric vector}
    \item{\code{X638}}{a numeric vector}
    \item{\code{X639}}{a numeric vector}
    \item{\code{X640}}{a numeric vector}
    \item{\code{X641}}{a numeric vector}
    \item{\code{X642}}{a numeric vector}
    \item{\code{X643}}{a numeric vector}
    \item{\code{X644}}{a numeric vector}
    \item{\code{X645}}{a numeric vector}
    \item{\code{X646}}{a numeric vector}
    \item{\code{X647}}{a numeric vector}
    \item{\code{X648}}{a numeric vector}
    \item{\code{X649}}{a numeric vector}
    \item{\code{X650}}{a numeric vector}
    \item{\code{X651}}{a numeric vector}
    \item{\code{X652}}{a numeric vector}
    \item{\code{X653}}{a numeric vector}
    \item{\code{X654}}{a numeric vector}
    \item{\code{X655}}{a numeric vector}
    \item{\code{X656}}{a numeric vector}
    \item{\code{X657}}{a numeric vector}
    \item{\code{X658}}{a numeric vector}
    \item{\code{X659}}{a numeric vector}
    \item{\code{X660}}{a numeric vector}
    \item{\code{X661}}{a numeric vector}
    \item{\code{X662}}{a numeric vector}
    \item{\code{X663}}{a numeric vector}
    \item{\code{X664}}{a numeric vector}
    \item{\code{X665}}{a numeric vector}
    \item{\code{X666}}{a numeric vector}
    \item{\code{X667}}{a numeric vector}
    \item{\code{X668}}{a numeric vector}
    \item{\code{X669}}{a numeric vector}
    \item{\code{X670}}{a numeric vector}
    \item{\code{X671}}{a numeric vector}
    \item{\code{X672}}{a numeric vector}
    \item{\code{X673}}{a numeric vector}
    \item{\code{X674}}{a numeric vector}
    \item{\code{X675}}{a numeric vector}
    \item{\code{X676}}{a numeric vector}
    \item{\code{X677}}{a numeric vector}
    \item{\code{X678}}{a numeric vector}
    \item{\code{X679}}{a numeric vector}
    \item{\code{X680}}{a numeric vector}
    \item{\code{X681}}{a numeric vector}
    \item{\code{X682}}{a numeric vector}
    \item{\code{X683}}{a numeric vector}
    \item{\code{X684}}{a numeric vector}
    \item{\code{X685}}{a numeric vector}
    \item{\code{X686}}{a numeric vector}
    \item{\code{X687}}{a numeric vector}
    \item{\code{X688}}{a numeric vector}
    \item{\code{X689}}{a numeric vector}
    \item{\code{X690}}{a numeric vector}
    \item{\code{X691}}{a numeric vector}
    \item{\code{X692}}{a numeric vector}
    \item{\code{X693}}{a numeric vector}
    \item{\code{X694}}{a numeric vector}
    \item{\code{X695}}{a numeric vector}
    \item{\code{X696}}{a numeric vector}
    \item{\code{X697}}{a numeric vector}
    \item{\code{X698}}{a numeric vector}
    \item{\code{X699}}{a numeric vector}
    \item{\code{X700}}{a numeric vector}
    \item{\code{X701}}{a numeric vector}
    \item{\code{X702}}{a numeric vector}
    \item{\code{X703}}{a numeric vector}
    \item{\code{X704}}{a numeric vector}
    \item{\code{X705}}{a numeric vector}
    \item{\code{X706}}{a numeric vector}
    \item{\code{X707}}{a numeric vector}
    \item{\code{X708}}{a numeric vector}
    \item{\code{X709}}{a numeric vector}
    \item{\code{X710}}{a numeric vector}
    \item{\code{X711}}{a numeric vector}
    \item{\code{X712}}{a numeric vector}
    \item{\code{X713}}{a numeric vector}
    \item{\code{X714}}{a numeric vector}
    \item{\code{X715}}{a numeric vector}
    \item{\code{X716}}{a numeric vector}
    \item{\code{X717}}{a numeric vector}
    \item{\code{X718}}{a numeric vector}
    \item{\code{X719}}{a numeric vector}
    \item{\code{X720}}{a numeric vector}
    \item{\code{X721}}{a numeric vector}
    \item{\code{X722}}{a numeric vector}
    \item{\code{X723}}{a numeric vector}
    \item{\code{X724}}{a numeric vector}
    \item{\code{X725}}{a numeric vector}
    \item{\code{X726}}{a numeric vector}
    \item{\code{X727}}{a numeric vector}
    \item{\code{X728}}{a numeric vector}
    \item{\code{X729}}{a numeric vector}
    \item{\code{X730}}{a numeric vector}
    \item{\code{X731}}{a numeric vector}
    \item{\code{X732}}{a numeric vector}
    \item{\code{X733}}{a numeric vector}
    \item{\code{X734}}{a numeric vector}
    \item{\code{X735}}{a numeric vector}
    \item{\code{X736}}{a numeric vector}
    \item{\code{X737}}{a numeric vector}
    \item{\code{X738}}{a numeric vector}
    \item{\code{X739}}{a numeric vector}
    \item{\code{X740}}{a numeric vector}
    \item{\code{X741}}{a numeric vector}
    \item{\code{X742}}{a numeric vector}
    \item{\code{X743}}{a numeric vector}
    \item{\code{X744}}{a numeric vector}
    \item{\code{X745}}{a numeric vector}
    \item{\code{X746}}{a numeric vector}
    \item{\code{X747}}{a numeric vector}
    \item{\code{X748}}{a numeric vector}
    \item{\code{X749}}{a numeric vector}
    \item{\code{X750}}{a numeric vector}
    \item{\code{X751}}{a numeric vector}
    \item{\code{X752}}{a numeric vector}
    \item{\code{X753}}{a numeric vector}
    \item{\code{X754}}{a numeric vector}
    \item{\code{X755}}{a numeric vector}
    \item{\code{X756}}{a numeric vector}
    \item{\code{X757}}{a numeric vector}
    \item{\code{X758}}{a numeric vector}
    \item{\code{X759}}{a numeric vector}
    \item{\code{X760}}{a numeric vector}
    \item{\code{X761}}{a numeric vector}
    \item{\code{X762}}{a numeric vector}
    \item{\code{X763}}{a numeric vector}
    \item{\code{X764}}{a numeric vector}
    \item{\code{X765}}{a numeric vector}
    \item{\code{X766}}{a numeric vector}
    \item{\code{X767}}{a numeric vector}
    \item{\code{X768}}{a numeric vector}
    \item{\code{X769}}{a numeric vector}
    \item{\code{X770}}{a numeric vector}
    \item{\code{X771}}{a numeric vector}
    \item{\code{X772}}{a numeric vector}
    \item{\code{X773}}{a numeric vector}
    \item{\code{X774}}{a numeric vector}
    \item{\code{X775}}{a numeric vector}
    \item{\code{X776}}{a numeric vector}
    \item{\code{X777}}{a numeric vector}
    \item{\code{X778}}{a numeric vector}
    \item{\code{X779}}{a numeric vector}
    \item{\code{X780}}{a numeric vector}
    \item{\code{X781}}{a numeric vector}
    \item{\code{X782}}{a numeric vector}
    \item{\code{X783}}{a numeric vector}
    \item{\code{X784}}{a numeric vector}
    \item{\code{X785}}{a numeric vector}
    \item{\code{X786}}{a numeric vector}
    \item{\code{X787}}{a numeric vector}
    \item{\code{X788}}{a numeric vector}
    \item{\code{X789}}{a numeric vector}
    \item{\code{X790}}{a numeric vector}
    \item{\code{X791}}{a numeric vector}
    \item{\code{X792}}{a numeric vector}
    \item{\code{X793}}{a numeric vector}
    \item{\code{X794}}{a numeric vector}
    \item{\code{X795}}{a numeric vector}
    \item{\code{X796}}{a numeric vector}
    \item{\code{X797}}{a numeric vector}
    \item{\code{X798}}{a numeric vector}
    \item{\code{X799}}{a numeric vector}
    \item{\code{X800}}{a numeric vector}
    \item{\code{X801}}{a numeric vector}
    \item{\code{X802}}{a numeric vector}
    \item{\code{X803}}{a numeric vector}
    \item{\code{X804}}{a numeric vector}
    \item{\code{X805}}{a numeric vector}
    \item{\code{X806}}{a numeric vector}
    \item{\code{X807}}{a numeric vector}
    \item{\code{X808}}{a numeric vector}
    \item{\code{X809}}{a numeric vector}
    \item{\code{X810}}{a numeric vector}
    \item{\code{X811}}{a numeric vector}
    \item{\code{X812}}{a numeric vector}
    \item{\code{X813}}{a numeric vector}
    \item{\code{X814}}{a numeric vector}
    \item{\code{X815}}{a numeric vector}
    \item{\code{X816}}{a numeric vector}
    \item{\code{X817}}{a numeric vector}
    \item{\code{X818}}{a numeric vector}
    \item{\code{X819}}{a numeric vector}
    \item{\code{X820}}{a numeric vector}
    \item{\code{X821}}{a numeric vector}
    \item{\code{X822}}{a numeric vector}
    \item{\code{X823}}{a numeric vector}
    \item{\code{X824}}{a numeric vector}
    \item{\code{X825}}{a numeric vector}
    \item{\code{X826}}{a numeric vector}
    \item{\code{X827}}{a numeric vector}
    \item{\code{X828}}{a numeric vector}
    \item{\code{X829}}{a numeric vector}
    \item{\code{X830}}{a numeric vector}
    \item{\code{X831}}{a numeric vector}
    \item{\code{X832}}{a numeric vector}
    \item{\code{X833}}{a numeric vector}
    \item{\code{X834}}{a numeric vector}
    \item{\code{X835}}{a numeric vector}
    \item{\code{X836}}{a numeric vector}
    \item{\code{X837}}{a numeric vector}
    \item{\code{X838}}{a numeric vector}
    \item{\code{X839}}{a numeric vector}
    \item{\code{X840}}{a numeric vector}
    \item{\code{X841}}{a numeric vector}
    \item{\code{X842}}{a numeric vector}
    \item{\code{X843}}{a numeric vector}
    \item{\code{X844}}{a numeric vector}
    \item{\code{X845}}{a numeric vector}
    \item{\code{X846}}{a numeric vector}
    \item{\code{X847}}{a numeric vector}
    \item{\code{X848}}{a numeric vector}
    \item{\code{X849}}{a numeric vector}
    \item{\code{X850}}{a numeric vector}
    \item{\code{X851}}{a numeric vector}
    \item{\code{X852}}{a numeric vector}
    \item{\code{X853}}{a numeric vector}
    \item{\code{X854}}{a numeric vector}
    \item{\code{X855}}{a numeric vector}
    \item{\code{X856}}{a numeric vector}
    \item{\code{X857}}{a numeric vector}
    \item{\code{X858}}{a numeric vector}
    \item{\code{X859}}{a numeric vector}
    \item{\code{X860}}{a numeric vector}
    \item{\code{X861}}{a numeric vector}
    \item{\code{X862}}{a numeric vector}
    \item{\code{X863}}{a numeric vector}
    \item{\code{X864}}{a numeric vector}
    \item{\code{X865}}{a numeric vector}
    \item{\code{X866}}{a numeric vector}
    \item{\code{X867}}{a numeric vector}
    \item{\code{X868}}{a numeric vector}
    \item{\code{X869}}{a numeric vector}
    \item{\code{X870}}{a numeric vector}
    \item{\code{X871}}{a numeric vector}
    \item{\code{X872}}{a numeric vector}
    \item{\code{X873}}{a numeric vector}
    \item{\code{X874}}{a numeric vector}
    \item{\code{X875}}{a numeric vector}
    \item{\code{X876}}{a numeric vector}
    \item{\code{X877}}{a numeric vector}
    \item{\code{X878}}{a numeric vector}
    \item{\code{X879}}{a numeric vector}
    \item{\code{X880}}{a numeric vector}
    \item{\code{X881}}{a numeric vector}
    \item{\code{X882}}{a numeric vector}
    \item{\code{X883}}{a numeric vector}
    \item{\code{X884}}{a numeric vector}
    \item{\code{X885}}{a numeric vector}
    \item{\code{X886}}{a numeric vector}
    \item{\code{X887}}{a numeric vector}
    \item{\code{X888}}{a numeric vector}
    \item{\code{X889}}{a numeric vector}
    \item{\code{X890}}{a numeric vector}
    \item{\code{X891}}{a numeric vector}
    \item{\code{X892}}{a numeric vector}
    \item{\code{X893}}{a numeric vector}
    \item{\code{X894}}{a numeric vector}
    \item{\code{X895}}{a numeric vector}
    \item{\code{X896}}{a numeric vector}
    \item{\code{X897}}{a numeric vector}
    \item{\code{X898}}{a numeric vector}
    \item{\code{X899}}{a numeric vector}
    \item{\code{X900}}{a numeric vector}
    \item{\code{X901}}{a numeric vector}
    \item{\code{X902}}{a numeric vector}
    \item{\code{X903}}{a numeric vector}
    \item{\code{X904}}{a numeric vector}
    \item{\code{X905}}{a numeric vector}
    \item{\code{X906}}{a numeric vector}
    \item{\code{X907}}{a numeric vector}
    \item{\code{X908}}{a numeric vector}
    \item{\code{X909}}{a numeric vector}
    \item{\code{X910}}{a numeric vector}
    \item{\code{X911}}{a numeric vector}
    \item{\code{X912}}{a numeric vector}
    \item{\code{X913}}{a numeric vector}
    \item{\code{X914}}{a numeric vector}
    \item{\code{X915}}{a numeric vector}
    \item{\code{X916}}{a numeric vector}
    \item{\code{X917}}{a numeric vector}
    \item{\code{X918}}{a numeric vector}
    \item{\code{X919}}{a numeric vector}
    \item{\code{X920}}{a numeric vector}
    \item{\code{X921}}{a numeric vector}
    \item{\code{X922}}{a numeric vector}
    \item{\code{X923}}{a numeric vector}
    \item{\code{X924}}{a numeric vector}
    \item{\code{X925}}{a numeric vector}
    \item{\code{X926}}{a numeric vector}
    \item{\code{X927}}{a numeric vector}
    \item{\code{X928}}{a numeric vector}
    \item{\code{X929}}{a numeric vector}
    \item{\code{X930}}{a numeric vector}
    \item{\code{X931}}{a numeric vector}
    \item{\code{X932}}{a numeric vector}
    \item{\code{X933}}{a numeric vector}
    \item{\code{X934}}{a numeric vector}
    \item{\code{X935}}{a numeric vector}
    \item{\code{X936}}{a numeric vector}
    \item{\code{X937}}{a numeric vector}
    \item{\code{X938}}{a numeric vector}
    \item{\code{X939}}{a numeric vector}
    \item{\code{X940}}{a numeric vector}
    \item{\code{X941}}{a numeric vector}
    \item{\code{X942}}{a numeric vector}
    \item{\code{X943}}{a numeric vector}
    \item{\code{X944}}{a numeric vector}
    \item{\code{X945}}{a numeric vector}
    \item{\code{X946}}{a numeric vector}
    \item{\code{X947}}{a numeric vector}
    \item{\code{X948}}{a numeric vector}
    \item{\code{X949}}{a numeric vector}
    \item{\code{X950}}{a numeric vector}
    \item{\code{X951}}{a numeric vector}
    \item{\code{X952}}{a numeric vector}
    \item{\code{X953}}{a numeric vector}
    \item{\code{X954}}{a numeric vector}
    \item{\code{X955}}{a numeric vector}
    \item{\code{X956}}{a numeric vector}
    \item{\code{X957}}{a numeric vector}
    \item{\code{X958}}{a numeric vector}
    \item{\code{X959}}{a numeric vector}
    \item{\code{X960}}{a numeric vector}
    \item{\code{X961}}{a numeric vector}
    \item{\code{X962}}{a numeric vector}
    \item{\code{X963}}{a numeric vector}
    \item{\code{X964}}{a numeric vector}
    \item{\code{X965}}{a numeric vector}
    \item{\code{X966}}{a numeric vector}
    \item{\code{X967}}{a numeric vector}
    \item{\code{X968}}{a numeric vector}
    \item{\code{X969}}{a numeric vector}
    \item{\code{X970}}{a numeric vector}
    \item{\code{X971}}{a numeric vector}
    \item{\code{X972}}{a numeric vector}
    \item{\code{X973}}{a numeric vector}
    \item{\code{X974}}{a numeric vector}
    \item{\code{X975}}{a numeric vector}
    \item{\code{X976}}{a numeric vector}
    \item{\code{X977}}{a numeric vector}
    \item{\code{X978}}{a numeric vector}
    \item{\code{X979}}{a numeric vector}
    \item{\code{X980}}{a numeric vector}
    \item{\code{X981}}{a numeric vector}
    \item{\code{X982}}{a numeric vector}
    \item{\code{X983}}{a numeric vector}
    \item{\code{X984}}{a numeric vector}
    \item{\code{X985}}{a numeric vector}
    \item{\code{X986}}{a numeric vector}
    \item{\code{X987}}{a numeric vector}
    \item{\code{X988}}{a numeric vector}
    \item{\code{X989}}{a numeric vector}
    \item{\code{X990}}{a numeric vector}
    \item{\code{X991}}{a numeric vector}
    \item{\code{X992}}{a numeric vector}
    \item{\code{X993}}{a numeric vector}
    \item{\code{X994}}{a numeric vector}
    \item{\code{X995}}{a numeric vector}
    \item{\code{X996}}{a numeric vector}
    \item{\code{X997}}{a numeric vector}
    \item{\code{X998}}{a numeric vector}
    \item{\code{X999}}{a numeric vector}
    \item{\code{X1000}}{a numeric vector}
    \item{\code{X1001}}{a numeric vector}
    \item{\code{X1002}}{a numeric vector}
    \item{\code{X1003}}{a numeric vector}
    \item{\code{X1004}}{a numeric vector}
    \item{\code{X1005}}{a numeric vector}
    \item{\code{X1006}}{a numeric vector}
    \item{\code{X1007}}{a numeric vector}
    \item{\code{X1008}}{a numeric vector}
    \item{\code{X1009}}{a numeric vector}
    \item{\code{X1010}}{a numeric vector}
    \item{\code{X1011}}{a numeric vector}
    \item{\code{X1012}}{a numeric vector}
    \item{\code{X1013}}{a numeric vector}
    \item{\code{X1014}}{a numeric vector}
    \item{\code{X1015}}{a numeric vector}
    \item{\code{X1016}}{a numeric vector}
    \item{\code{X1017}}{a numeric vector}
    \item{\code{X1018}}{a numeric vector}
    \item{\code{X1019}}{a numeric vector}
    \item{\code{X1020}}{a numeric vector}
    \item{\code{X1021}}{a numeric vector}
    \item{\code{X1022}}{a numeric vector}
    \item{\code{X1023}}{a numeric vector}
    \item{\code{X1024}}{a numeric vector}
    \item{\code{X1025}}{a numeric vector}
    \item{\code{X1026}}{a numeric vector}
    \item{\code{X1027}}{a numeric vector}
    \item{\code{X1028}}{a numeric vector}
    \item{\code{X1029}}{a numeric vector}
    \item{\code{X1030}}{a numeric vector}
    \item{\code{X1031}}{a numeric vector}
    \item{\code{X1032}}{a numeric vector}
    \item{\code{X1033}}{a numeric vector}
    \item{\code{X1034}}{a numeric vector}
    \item{\code{X1035}}{a numeric vector}
    \item{\code{X1036}}{a numeric vector}
    \item{\code{X1037}}{a numeric vector}
    \item{\code{X1038}}{a numeric vector}
    \item{\code{X1039}}{a numeric vector}
    \item{\code{X1040}}{a numeric vector}
    \item{\code{X1041}}{a numeric vector}
    \item{\code{X1042}}{a numeric vector}
    \item{\code{X1043}}{a numeric vector}
    \item{\code{X1044}}{a numeric vector}
    \item{\code{X1045}}{a numeric vector}
    \item{\code{X1046}}{a numeric vector}
    \item{\code{X1047}}{a numeric vector}
    \item{\code{X1048}}{a numeric vector}
    \item{\code{X1049}}{a numeric vector}
    \item{\code{X1050}}{a numeric vector}
    \item{\code{X1051}}{a numeric vector}
    \item{\code{X1052}}{a numeric vector}
    \item{\code{X1053}}{a numeric vector}
    \item{\code{X1054}}{a numeric vector}
    \item{\code{X1055}}{a numeric vector}
    \item{\code{X1056}}{a numeric vector}
    \item{\code{X1057}}{a numeric vector}
    \item{\code{X1058}}{a numeric vector}
    \item{\code{X1059}}{a numeric vector}
    \item{\code{X1060}}{a numeric vector}
    \item{\code{X1061}}{a numeric vector}
    \item{\code{X1062}}{a numeric vector}
    \item{\code{X1063}}{a numeric vector}
    \item{\code{X1064}}{a numeric vector}
    \item{\code{X1065}}{a numeric vector}
    \item{\code{X1066}}{a numeric vector}
    \item{\code{X1067}}{a numeric vector}
    \item{\code{X1068}}{a numeric vector}
    \item{\code{X1069}}{a numeric vector}
    \item{\code{X1070}}{a numeric vector}
    \item{\code{X1071}}{a numeric vector}
    \item{\code{X1072}}{a numeric vector}
    \item{\code{X1073}}{a numeric vector}
    \item{\code{X1074}}{a numeric vector}
    \item{\code{X1075}}{a numeric vector}
    \item{\code{X1076}}{a numeric vector}
    \item{\code{X1077}}{a numeric vector}
    \item{\code{X1078}}{a numeric vector}
    \item{\code{X1079}}{a numeric vector}
    \item{\code{X1080}}{a numeric vector}
    \item{\code{X1081}}{a numeric vector}
    \item{\code{X1082}}{a numeric vector}
    \item{\code{X1083}}{a numeric vector}
    \item{\code{X1084}}{a numeric vector}
    \item{\code{X1085}}{a numeric vector}
    \item{\code{X1086}}{a numeric vector}
    \item{\code{X1087}}{a numeric vector}
    \item{\code{X1088}}{a numeric vector}
    \item{\code{X1089}}{a numeric vector}
    \item{\code{X1090}}{a numeric vector}
    \item{\code{X1091}}{a numeric vector}
    \item{\code{X1092}}{a numeric vector}
    \item{\code{X1093}}{a numeric vector}
    \item{\code{X1094}}{a numeric vector}
    \item{\code{X1095}}{a numeric vector}
    \item{\code{X1096}}{a numeric vector}
    \item{\code{X1097}}{a numeric vector}
    \item{\code{X1098}}{a numeric vector}
    \item{\code{X1099}}{a numeric vector}
    \item{\code{X1100}}{a numeric vector}
    \item{\code{X1101}}{a numeric vector}
    \item{\code{X1102}}{a numeric vector}
    \item{\code{X1103}}{a numeric vector}
    \item{\code{X1104}}{a numeric vector}
    \item{\code{X1105}}{a numeric vector}
    \item{\code{X1106}}{a numeric vector}
    \item{\code{X1107}}{a numeric vector}
    \item{\code{X1108}}{a numeric vector}
    \item{\code{X1109}}{a numeric vector}
    \item{\code{X1110}}{a numeric vector}
    \item{\code{X1111}}{a numeric vector}
    \item{\code{X1112}}{a numeric vector}
    \item{\code{X1113}}{a numeric vector}
    \item{\code{X1114}}{a numeric vector}
    \item{\code{X1115}}{a numeric vector}
    \item{\code{X1116}}{a numeric vector}
    \item{\code{X1117}}{a numeric vector}
    \item{\code{X1118}}{a numeric vector}
    \item{\code{X1119}}{a numeric vector}
    \item{\code{X1120}}{a numeric vector}
    \item{\code{X1121}}{a numeric vector}
    \item{\code{X1122}}{a numeric vector}
    \item{\code{X1123}}{a numeric vector}
    \item{\code{X1124}}{a numeric vector}
    \item{\code{X1125}}{a numeric vector}
    \item{\code{X1126}}{a numeric vector}
    \item{\code{X1127}}{a numeric vector}
    \item{\code{X1128}}{a numeric vector}
    \item{\code{X1129}}{a numeric vector}
    \item{\code{X1130}}{a numeric vector}
    \item{\code{X1131}}{a numeric vector}
    \item{\code{X1132}}{a numeric vector}
    \item{\code{X1133}}{a numeric vector}
    \item{\code{X1134}}{a numeric vector}
    \item{\code{X1135}}{a numeric vector}
    \item{\code{X1136}}{a numeric vector}
    \item{\code{X1137}}{a numeric vector}
    \item{\code{X1138}}{a numeric vector}
    \item{\code{X1139}}{a numeric vector}
    \item{\code{X1140}}{a numeric vector}
    \item{\code{X1141}}{a numeric vector}
    \item{\code{X1142}}{a numeric vector}
    \item{\code{X1143}}{a numeric vector}
    \item{\code{X1144}}{a numeric vector}
    \item{\code{X1145}}{a numeric vector}
    \item{\code{X1146}}{a numeric vector}
    \item{\code{X1147}}{a numeric vector}
    \item{\code{X1148}}{a numeric vector}
    \item{\code{X1149}}{a numeric vector}
    \item{\code{X1150}}{a numeric vector}
    \item{\code{X1151}}{a numeric vector}
    \item{\code{X1152}}{a numeric vector}
    \item{\code{X1153}}{a numeric vector}
    \item{\code{X1154}}{a numeric vector}
    \item{\code{X1155}}{a numeric vector}
    \item{\code{X1156}}{a numeric vector}
    \item{\code{X1157}}{a numeric vector}
    \item{\code{X1158}}{a numeric vector}
    \item{\code{X1159}}{a numeric vector}
    \item{\code{X1160}}{a numeric vector}
    \item{\code{X1161}}{a numeric vector}
    \item{\code{X1162}}{a numeric vector}
    \item{\code{X1163}}{a numeric vector}
    \item{\code{X1164}}{a numeric vector}
    \item{\code{X1165}}{a numeric vector}
    \item{\code{X1166}}{a numeric vector}
    \item{\code{X1167}}{a numeric vector}
    \item{\code{X1168}}{a numeric vector}
    \item{\code{X1169}}{a numeric vector}
    \item{\code{X1170}}{a numeric vector}
    \item{\code{X1171}}{a numeric vector}
    \item{\code{X1172}}{a numeric vector}
    \item{\code{X1173}}{a numeric vector}
    \item{\code{X1174}}{a numeric vector}
    \item{\code{X1175}}{a numeric vector}
    \item{\code{X1176}}{a numeric vector}
    \item{\code{X1177}}{a numeric vector}
    \item{\code{X1178}}{a numeric vector}
    \item{\code{X1179}}{a numeric vector}
    \item{\code{X1180}}{a numeric vector}
    \item{\code{X1181}}{a numeric vector}
    \item{\code{X1182}}{a numeric vector}
    \item{\code{X1183}}{a numeric vector}
    \item{\code{X1184}}{a numeric vector}
    \item{\code{X1185}}{a numeric vector}
    \item{\code{X1186}}{a numeric vector}
    \item{\code{X1187}}{a numeric vector}
    \item{\code{X1188}}{a numeric vector}
    \item{\code{X1189}}{a numeric vector}
    \item{\code{X1190}}{a numeric vector}
    \item{\code{X1191}}{a numeric vector}
    \item{\code{X1192}}{a numeric vector}
    \item{\code{X1193}}{a numeric vector}
    \item{\code{X1194}}{a numeric vector}
    \item{\code{X1195}}{a numeric vector}
    \item{\code{X1196}}{a numeric vector}
    \item{\code{X1197}}{a numeric vector}
    \item{\code{X1198}}{a numeric vector}
    \item{\code{X1199}}{a numeric vector}
    \item{\code{X1200}}{a numeric vector}
    \item{\code{X1201}}{a numeric vector}
    \item{\code{X1202}}{a numeric vector}
    \item{\code{X1203}}{a numeric vector}
    \item{\code{X1204}}{a numeric vector}
    \item{\code{X1205}}{a numeric vector}
    \item{\code{X1206}}{a numeric vector}
    \item{\code{X1207}}{a numeric vector}
    \item{\code{X1208}}{a numeric vector}
    \item{\code{X1209}}{a numeric vector}
    \item{\code{X1210}}{a numeric vector}
    \item{\code{X1211}}{a numeric vector}
    \item{\code{X1212}}{a numeric vector}
    \item{\code{X1213}}{a numeric vector}
    \item{\code{X1214}}{a numeric vector}
    \item{\code{X1215}}{a numeric vector}
    \item{\code{X1216}}{a numeric vector}
    \item{\code{X1217}}{a numeric vector}
    \item{\code{X1218}}{a numeric vector}
    \item{\code{X1219}}{a numeric vector}
    \item{\code{X1220}}{a numeric vector}
    \item{\code{X1221}}{a numeric vector}
    \item{\code{X1222}}{a numeric vector}
    \item{\code{X1223}}{a numeric vector}
    \item{\code{X1224}}{a numeric vector}
    \item{\code{X1225}}{a numeric vector}
    \item{\code{X1226}}{a numeric vector}
    \item{\code{X1227}}{a numeric vector}
    \item{\code{X1228}}{a numeric vector}
    \item{\code{X1229}}{a numeric vector}
    \item{\code{X1230}}{a numeric vector}
    \item{\code{X1231}}{a numeric vector}
    \item{\code{X1232}}{a numeric vector}
    \item{\code{X1233}}{a numeric vector}
    \item{\code{X1234}}{a numeric vector}
    \item{\code{X1235}}{a numeric vector}
    \item{\code{X1236}}{a numeric vector}
    \item{\code{X1237}}{a numeric vector}
    \item{\code{X1238}}{a numeric vector}
    \item{\code{X1239}}{a numeric vector}
    \item{\code{X1240}}{a numeric vector}
    \item{\code{X1241}}{a numeric vector}
    \item{\code{X1242}}{a numeric vector}
    \item{\code{X1243}}{a numeric vector}
    \item{\code{X1244}}{a numeric vector}
    \item{\code{X1245}}{a numeric vector}
    \item{\code{X1246}}{a numeric vector}
    \item{\code{X1247}}{a numeric vector}
    \item{\code{X1248}}{a numeric vector}
    \item{\code{X1249}}{a numeric vector}
    \item{\code{X1250}}{a numeric vector}
    \item{\code{X1251}}{a numeric vector}
    \item{\code{X1252}}{a numeric vector}
    \item{\code{X1253}}{a numeric vector}
    \item{\code{X1254}}{a numeric vector}
    \item{\code{X1255}}{a numeric vector}
    \item{\code{X1256}}{a numeric vector}
    \item{\code{X1257}}{a numeric vector}
    \item{\code{X1258}}{a numeric vector}
    \item{\code{X1259}}{a numeric vector}
    \item{\code{X1260}}{a numeric vector}
    \item{\code{X1261}}{a numeric vector}
    \item{\code{X1262}}{a numeric vector}
    \item{\code{X1263}}{a numeric vector}
    \item{\code{X1264}}{a numeric vector}
    \item{\code{X1265}}{a numeric vector}
    \item{\code{X1266}}{a numeric vector}
    \item{\code{X1267}}{a numeric vector}
    \item{\code{X1268}}{a numeric vector}
    \item{\code{X1269}}{a numeric vector}
    \item{\code{X1270}}{a numeric vector}
    \item{\code{X1271}}{a numeric vector}
    \item{\code{X1272}}{a numeric vector}
    \item{\code{X1273}}{a numeric vector}
    \item{\code{X1274}}{a numeric vector}
    \item{\code{X1275}}{a numeric vector}
    \item{\code{X1276}}{a numeric vector}
    \item{\code{X1277}}{a numeric vector}
    \item{\code{X1278}}{a numeric vector}
    \item{\code{X1279}}{a numeric vector}
    \item{\code{X1280}}{a numeric vector}
    \item{\code{X1281}}{a numeric vector}
    \item{\code{X1282}}{a numeric vector}
    \item{\code{X1283}}{a numeric vector}
    \item{\code{X1284}}{a numeric vector}
    \item{\code{X1285}}{a numeric vector}
    \item{\code{X1286}}{a numeric vector}
    \item{\code{X1287}}{a numeric vector}
    \item{\code{X1288}}{a numeric vector}
    \item{\code{X1289}}{a numeric vector}
    \item{\code{X1290}}{a numeric vector}
    \item{\code{X1291}}{a numeric vector}
    \item{\code{X1292}}{a numeric vector}
    \item{\code{X1293}}{a numeric vector}
    \item{\code{X1294}}{a numeric vector}
    \item{\code{X1295}}{a numeric vector}
    \item{\code{X1296}}{a numeric vector}
    \item{\code{X1297}}{a numeric vector}
    \item{\code{X1298}}{a numeric vector}
    \item{\code{X1299}}{a numeric vector}
    \item{\code{X1300}}{a numeric vector}
    \item{\code{X1301}}{a numeric vector}
    \item{\code{X1302}}{a numeric vector}
    \item{\code{X1303}}{a numeric vector}
    \item{\code{X1304}}{a numeric vector}
    \item{\code{X1305}}{a numeric vector}
    \item{\code{X1306}}{a numeric vector}
    \item{\code{X1307}}{a numeric vector}
    \item{\code{X1308}}{a numeric vector}
    \item{\code{X1309}}{a numeric vector}
    \item{\code{X1310}}{a numeric vector}
    \item{\code{X1311}}{a numeric vector}
    \item{\code{X1312}}{a numeric vector}
    \item{\code{X1313}}{a numeric vector}
    \item{\code{X1314}}{a numeric vector}
    \item{\code{X1315}}{a numeric vector}
    \item{\code{X1316}}{a numeric vector}
    \item{\code{X1317}}{a numeric vector}
    \item{\code{X1318}}{a numeric vector}
    \item{\code{X1319}}{a numeric vector}
    \item{\code{X1320}}{a numeric vector}
    \item{\code{X1321}}{a numeric vector}
    \item{\code{X1322}}{a numeric vector}
    \item{\code{X1323}}{a numeric vector}
    \item{\code{X1324}}{a numeric vector}
    \item{\code{X1325}}{a numeric vector}
    \item{\code{X1326}}{a numeric vector}
    \item{\code{X1327}}{a numeric vector}
    \item{\code{X1328}}{a numeric vector}
    \item{\code{X1329}}{a numeric vector}
    \item{\code{X1330}}{a numeric vector}
    \item{\code{X1331}}{a numeric vector}
    \item{\code{X1332}}{a numeric vector}
    \item{\code{X1333}}{a numeric vector}
    \item{\code{X1334}}{a numeric vector}
    \item{\code{X1335}}{a numeric vector}
    \item{\code{X1336}}{a numeric vector}
    \item{\code{X1337}}{a numeric vector}
    \item{\code{X1338}}{a numeric vector}
    \item{\code{X1339}}{a numeric vector}
    \item{\code{X1340}}{a numeric vector}
    \item{\code{X1341}}{a numeric vector}
    \item{\code{X1342}}{a numeric vector}
    \item{\code{X1343}}{a numeric vector}
    \item{\code{X1344}}{a numeric vector}
    \item{\code{X1345}}{a numeric vector}
    \item{\code{X1346}}{a numeric vector}
    \item{\code{X1347}}{a numeric vector}
    \item{\code{X1348}}{a numeric vector}
    \item{\code{X1349}}{a numeric vector}
    \item{\code{X1350}}{a numeric vector}
    \item{\code{X1351}}{a numeric vector}
    \item{\code{X1352}}{a numeric vector}
    \item{\code{X1353}}{a numeric vector}
    \item{\code{X1354}}{a numeric vector}
    \item{\code{X1355}}{a numeric vector}
    \item{\code{X1356}}{a numeric vector}
    \item{\code{X1357}}{a numeric vector}
    \item{\code{X1358}}{a numeric vector}
    \item{\code{X1359}}{a numeric vector}
    \item{\code{X1360}}{a numeric vector}
    \item{\code{X1361}}{a numeric vector}
    \item{\code{X1362}}{a numeric vector}
    \item{\code{X1363}}{a numeric vector}
    \item{\code{X1364}}{a numeric vector}
    \item{\code{X1365}}{a numeric vector}
    \item{\code{X1366}}{a numeric vector}
    \item{\code{X1367}}{a numeric vector}
    \item{\code{X1368}}{a numeric vector}
    \item{\code{X1369}}{a numeric vector}
    \item{\code{X1370}}{a numeric vector}
    \item{\code{X1371}}{a numeric vector}
    \item{\code{X1372}}{a numeric vector}
    \item{\code{X1373}}{a numeric vector}
    \item{\code{X1374}}{a numeric vector}
    \item{\code{X1375}}{a numeric vector}
    \item{\code{X1376}}{a numeric vector}
    \item{\code{X1377}}{a numeric vector}
    \item{\code{X1378}}{a numeric vector}
    \item{\code{X1379}}{a numeric vector}
    \item{\code{X1380}}{a numeric vector}
    \item{\code{X1381}}{a numeric vector}
    \item{\code{X1382}}{a numeric vector}
    \item{\code{X1383}}{a numeric vector}
    \item{\code{X1384}}{a numeric vector}
    \item{\code{X1385}}{a numeric vector}
    \item{\code{X1386}}{a numeric vector}
    \item{\code{X1387}}{a numeric vector}
    \item{\code{X1388}}{a numeric vector}
    \item{\code{X1389}}{a numeric vector}
    \item{\code{X1390}}{a numeric vector}
    \item{\code{X1391}}{a numeric vector}
    \item{\code{X1392}}{a numeric vector}
    \item{\code{X1393}}{a numeric vector}
    \item{\code{X1394}}{a numeric vector}
    \item{\code{X1395}}{a numeric vector}
    \item{\code{X1396}}{a numeric vector}
    \item{\code{X1397}}{a numeric vector}
    \item{\code{X1398}}{a numeric vector}
    \item{\code{X1399}}{a numeric vector}
    \item{\code{X1400}}{a numeric vector}
    \item{\code{X1401}}{a numeric vector}
    \item{\code{X1402}}{a numeric vector}
    \item{\code{X1403}}{a numeric vector}
    \item{\code{X1404}}{a numeric vector}
    \item{\code{X1405}}{a numeric vector}
    \item{\code{X1406}}{a numeric vector}
    \item{\code{X1407}}{a numeric vector}
    \item{\code{X1408}}{a numeric vector}
    \item{\code{X1409}}{a numeric vector}
    \item{\code{X1410}}{a numeric vector}
    \item{\code{X1411}}{a numeric vector}
    \item{\code{X1412}}{a numeric vector}
    \item{\code{X1413}}{a numeric vector}
    \item{\code{X1414}}{a numeric vector}
    \item{\code{X1415}}{a numeric vector}
    \item{\code{X1416}}{a numeric vector}
    \item{\code{X1417}}{a numeric vector}
    \item{\code{X1418}}{a numeric vector}
    \item{\code{X1419}}{a numeric vector}
    \item{\code{X1420}}{a numeric vector}
    \item{\code{X1421}}{a numeric vector}
    \item{\code{X1422}}{a numeric vector}
    \item{\code{X1423}}{a numeric vector}
    \item{\code{X1424}}{a numeric vector}
    \item{\code{X1425}}{a numeric vector}
    \item{\code{X1426}}{a numeric vector}
    \item{\code{X1427}}{a numeric vector}
    \item{\code{X1428}}{a numeric vector}
    \item{\code{X1429}}{a numeric vector}
    \item{\code{X1430}}{a numeric vector}
    \item{\code{X1431}}{a numeric vector}
    \item{\code{X1432}}{a numeric vector}
    \item{\code{X1433}}{a numeric vector}
    \item{\code{X1434}}{a numeric vector}
    \item{\code{X1435}}{a numeric vector}
    \item{\code{X1436}}{a numeric vector}
    \item{\code{X1437}}{a numeric vector}
    \item{\code{X1438}}{a numeric vector}
    \item{\code{X1439}}{a numeric vector}
    \item{\code{X1440}}{a numeric vector}
    \item{\code{X1441}}{a numeric vector}
    \item{\code{X1442}}{a numeric vector}
    \item{\code{X1443}}{a numeric vector}
    \item{\code{X1444}}{a numeric vector}
    \item{\code{X1445}}{a numeric vector}
    \item{\code{X1446}}{a numeric vector}
    \item{\code{X1447}}{a numeric vector}
    \item{\code{X1448}}{a numeric vector}
    \item{\code{X1449}}{a numeric vector}
    \item{\code{X1450}}{a numeric vector}
    \item{\code{X1451}}{a numeric vector}
    \item{\code{X1452}}{a numeric vector}
    \item{\code{X1453}}{a numeric vector}
    \item{\code{X1454}}{a numeric vector}
    \item{\code{X1455}}{a numeric vector}
    \item{\code{X1456}}{a numeric vector}
    \item{\code{X1457}}{a numeric vector}
    \item{\code{X1458}}{a numeric vector}
    \item{\code{X1459}}{a numeric vector}
    \item{\code{X1460}}{a numeric vector}
    \item{\code{X1461}}{a numeric vector}
    \item{\code{X1462}}{a numeric vector}
    \item{\code{X1463}}{a numeric vector}
    \item{\code{X1464}}{a numeric vector}
    \item{\code{X1465}}{a numeric vector}
    \item{\code{X1466}}{a numeric vector}
    \item{\code{X1467}}{a numeric vector}
    \item{\code{X1468}}{a numeric vector}
    \item{\code{X1469}}{a numeric vector}
    \item{\code{X1470}}{a numeric vector}
    \item{\code{X1471}}{a numeric vector}
    \item{\code{X1472}}{a numeric vector}
    \item{\code{X1473}}{a numeric vector}
    \item{\code{X1474}}{a numeric vector}
    \item{\code{X1475}}{a numeric vector}
    \item{\code{X1476}}{a numeric vector}
    \item{\code{X1477}}{a numeric vector}
    \item{\code{X1478}}{a numeric vector}
    \item{\code{X1479}}{a numeric vector}
    \item{\code{X1480}}{a numeric vector}
    \item{\code{X1481}}{a numeric vector}
    \item{\code{X1482}}{a numeric vector}
    \item{\code{X1483}}{a numeric vector}
    \item{\code{X1484}}{a numeric vector}
    \item{\code{X1485}}{a numeric vector}
    \item{\code{X1486}}{a numeric vector}
    \item{\code{X1487}}{a numeric vector}
    \item{\code{X1488}}{a numeric vector}
    \item{\code{X1489}}{a numeric vector}
    \item{\code{X1490}}{a numeric vector}
    \item{\code{X1491}}{a numeric vector}
    \item{\code{X1492}}{a numeric vector}
    \item{\code{X1493}}{a numeric vector}
    \item{\code{X1494}}{a numeric vector}
    \item{\code{X1495}}{a numeric vector}
    \item{\code{X1496}}{a numeric vector}
    \item{\code{X1497}}{a numeric vector}
    \item{\code{X1498}}{a numeric vector}
    \item{\code{X1499}}{a numeric vector}
    \item{\code{X1500}}{a numeric vector}
    \item{\code{X1501}}{a numeric vector}
    \item{\code{X1502}}{a numeric vector}
    \item{\code{X1503}}{a numeric vector}
    \item{\code{X1504}}{a numeric vector}
    \item{\code{X1505}}{a numeric vector}
    \item{\code{X1506}}{a numeric vector}
    \item{\code{X1507}}{a numeric vector}
    \item{\code{X1508}}{a numeric vector}
    \item{\code{X1509}}{a numeric vector}
    \item{\code{X1510}}{a numeric vector}
    \item{\code{X1511}}{a numeric vector}
    \item{\code{X1512}}{a numeric vector}
    \item{\code{X1513}}{a numeric vector}
    \item{\code{X1514}}{a numeric vector}
    \item{\code{X1515}}{a numeric vector}
    \item{\code{X1516}}{a numeric vector}
    \item{\code{X1517}}{a numeric vector}
    \item{\code{X1518}}{a numeric vector}
    \item{\code{X1519}}{a numeric vector}
    \item{\code{X1520}}{a numeric vector}
    \item{\code{X1521}}{a numeric vector}
    \item{\code{X1522}}{a numeric vector}
    \item{\code{X1523}}{a numeric vector}
    \item{\code{X1524}}{a numeric vector}
    \item{\code{X1525}}{a numeric vector}
    \item{\code{X1526}}{a numeric vector}
    \item{\code{X1527}}{a numeric vector}
    \item{\code{X1528}}{a numeric vector}
    \item{\code{X1529}}{a numeric vector}
    \item{\code{X1530}}{a numeric vector}
    \item{\code{X1531}}{a numeric vector}
    \item{\code{X1532}}{a numeric vector}
    \item{\code{X1533}}{a numeric vector}
    \item{\code{X1534}}{a numeric vector}
    \item{\code{X1535}}{a numeric vector}
    \item{\code{X1536}}{a numeric vector}
    \item{\code{X1537}}{a numeric vector}
    \item{\code{X1538}}{a numeric vector}
    \item{\code{X1539}}{a numeric vector}
    \item{\code{X1540}}{a numeric vector}
    \item{\code{X1541}}{a numeric vector}
    \item{\code{X1542}}{a numeric vector}
    \item{\code{X1543}}{a numeric vector}
    \item{\code{X1544}}{a numeric vector}
    \item{\code{X1545}}{a numeric vector}
    \item{\code{X1546}}{a numeric vector}
    \item{\code{X1547}}{a numeric vector}
    \item{\code{X1548}}{a numeric vector}
    \item{\code{X1549}}{a numeric vector}
    \item{\code{X1550}}{a numeric vector}
    \item{\code{X1551}}{a numeric vector}
    \item{\code{X1552}}{a numeric vector}
    \item{\code{X1553}}{a numeric vector}
    \item{\code{X1554}}{a numeric vector}
    \item{\code{X1555}}{a numeric vector}
    \item{\code{X1556}}{a numeric vector}
    \item{\code{X1557}}{a numeric vector}
    \item{\code{X1558}}{a numeric vector}
    \item{\code{X1559}}{a numeric vector}
    \item{\code{X1560}}{a numeric vector}
    \item{\code{X1561}}{a numeric vector}
    \item{\code{X1562}}{a numeric vector}
    \item{\code{X1563}}{a numeric vector}
    \item{\code{X1564}}{a numeric vector}
    \item{\code{X1565}}{a numeric vector}
    \item{\code{X1566}}{a numeric vector}
    \item{\code{X1567}}{a numeric vector}
    \item{\code{X1568}}{a numeric vector}
    \item{\code{X1569}}{a numeric vector}
    \item{\code{X1570}}{a numeric vector}
    \item{\code{X1571}}{a numeric vector}
    \item{\code{X1572}}{a numeric vector}
    \item{\code{X1573}}{a numeric vector}
    \item{\code{X1574}}{a numeric vector}
    \item{\code{X1575}}{a numeric vector}
    \item{\code{X1576}}{a numeric vector}
    \item{\code{X1577}}{a numeric vector}
    \item{\code{X1578}}{a numeric vector}
    \item{\code{X1579}}{a numeric vector}
    \item{\code{X1580}}{a numeric vector}
    \item{\code{X1581}}{a numeric vector}
    \item{\code{X1582}}{a numeric vector}
    \item{\code{X1583}}{a numeric vector}
    \item{\code{X1584}}{a numeric vector}
    \item{\code{X1585}}{a numeric vector}
    \item{\code{X1586}}{a numeric vector}
    \item{\code{X1587}}{a numeric vector}
    \item{\code{X1588}}{a numeric vector}
    \item{\code{X1589}}{a numeric vector}
    \item{\code{X1590}}{a numeric vector}
    \item{\code{X1591}}{a numeric vector}
    \item{\code{X1592}}{a numeric vector}
    \item{\code{X1593}}{a numeric vector}
    \item{\code{X1594}}{a numeric vector}
    \item{\code{X1595}}{a numeric vector}
    \item{\code{X1596}}{a numeric vector}
    \item{\code{X1597}}{a numeric vector}
    \item{\code{X1598}}{a numeric vector}
    \item{\code{X1599}}{a numeric vector}
    \item{\code{X1600}}{a numeric vector}
    \item{\code{X1601}}{a numeric vector}
    \item{\code{X1602}}{a numeric vector}
    \item{\code{X1603}}{a numeric vector}
    \item{\code{X1604}}{a numeric vector}
    \item{\code{X1605}}{a numeric vector}
    \item{\code{X1606}}{a numeric vector}
    \item{\code{X1607}}{a numeric vector}
    \item{\code{X1608}}{a numeric vector}
    \item{\code{X1609}}{a numeric vector}
    \item{\code{X1610}}{a numeric vector}
    \item{\code{X1611}}{a numeric vector}
    \item{\code{X1612}}{a numeric vector}
    \item{\code{X1613}}{a numeric vector}
    \item{\code{X1614}}{a numeric vector}
    \item{\code{X1615}}{a numeric vector}
    \item{\code{X1616}}{a numeric vector}
    \item{\code{X1617}}{a numeric vector}
    \item{\code{X1618}}{a numeric vector}
    \item{\code{X1619}}{a numeric vector}
    \item{\code{X1620}}{a numeric vector}
    \item{\code{X1621}}{a numeric vector}
    \item{\code{X1622}}{a numeric vector}
    \item{\code{X1623}}{a numeric vector}
    \item{\code{X1624}}{a numeric vector}
    \item{\code{X1625}}{a numeric vector}
    \item{\code{X1626}}{a numeric vector}
    \item{\code{X1627}}{a numeric vector}
    \item{\code{X1628}}{a numeric vector}
    \item{\code{X1629}}{a numeric vector}
    \item{\code{X1630}}{a numeric vector}
    \item{\code{X1631}}{a numeric vector}
    \item{\code{X1632}}{a numeric vector}
    \item{\code{X1633}}{a numeric vector}
    \item{\code{X1634}}{a numeric vector}
    \item{\code{X1635}}{a numeric vector}
    \item{\code{X1636}}{a numeric vector}
    \item{\code{X1637}}{a numeric vector}
    \item{\code{X1638}}{a numeric vector}
    \item{\code{X1639}}{a numeric vector}
    \item{\code{X1640}}{a numeric vector}
    \item{\code{X1641}}{a numeric vector}
    \item{\code{X1642}}{a numeric vector}
    \item{\code{X1643}}{a numeric vector}
    \item{\code{X1644}}{a numeric vector}
    \item{\code{X1645}}{a numeric vector}
    \item{\code{X1646}}{a numeric vector}
    \item{\code{X1647}}{a numeric vector}
    \item{\code{X1648}}{a numeric vector}
    \item{\code{X1649}}{a numeric vector}
    \item{\code{X1650}}{a numeric vector}
    \item{\code{X1651}}{a numeric vector}
    \item{\code{X1652}}{a numeric vector}
    \item{\code{X1653}}{a numeric vector}
    \item{\code{X1654}}{a numeric vector}
    \item{\code{X1655}}{a numeric vector}
    \item{\code{X1656}}{a numeric vector}
    \item{\code{X1657}}{a numeric vector}
    \item{\code{X1658}}{a numeric vector}
    \item{\code{X1659}}{a numeric vector}
    \item{\code{X1660}}{a numeric vector}
    \item{\code{X1661}}{a numeric vector}
    \item{\code{X1662}}{a numeric vector}
    \item{\code{X1663}}{a numeric vector}
    \item{\code{X1664}}{a numeric vector}
    \item{\code{X1665}}{a numeric vector}
    \item{\code{X1666}}{a numeric vector}
    \item{\code{X1667}}{a numeric vector}
    \item{\code{X1668}}{a numeric vector}
    \item{\code{X1669}}{a numeric vector}
    \item{\code{X1670}}{a numeric vector}
    \item{\code{X1671}}{a numeric vector}
    \item{\code{X1672}}{a numeric vector}
    \item{\code{X1673}}{a numeric vector}
    \item{\code{X1674}}{a numeric vector}
    \item{\code{X1675}}{a numeric vector}
    \item{\code{X1676}}{a numeric vector}
    \item{\code{X1677}}{a numeric vector}
    \item{\code{X1678}}{a numeric vector}
    \item{\code{X1679}}{a numeric vector}
    \item{\code{X1680}}{a numeric vector}
    \item{\code{X1681}}{a numeric vector}
    \item{\code{X1682}}{a numeric vector}
    \item{\code{X1683}}{a numeric vector}
    \item{\code{X1684}}{a numeric vector}
    \item{\code{X1685}}{a numeric vector}
    \item{\code{X1686}}{a numeric vector}
    \item{\code{X1687}}{a numeric vector}
    \item{\code{X1688}}{a numeric vector}
    \item{\code{X1689}}{a numeric vector}
    \item{\code{X1690}}{a numeric vector}
    \item{\code{X1691}}{a numeric vector}
    \item{\code{X1692}}{a numeric vector}
    \item{\code{X1693}}{a numeric vector}
    \item{\code{X1694}}{a numeric vector}
    \item{\code{X1695}}{a numeric vector}
    \item{\code{X1696}}{a numeric vector}
    \item{\code{X1697}}{a numeric vector}
    \item{\code{X1698}}{a numeric vector}
    \item{\code{X1699}}{a numeric vector}
    \item{\code{X1700}}{a numeric vector}
    \item{\code{X1701}}{a numeric vector}
    \item{\code{X1702}}{a numeric vector}
    \item{\code{X1703}}{a numeric vector}
    \item{\code{X1704}}{a numeric vector}
    \item{\code{X1705}}{a numeric vector}
    \item{\code{X1706}}{a numeric vector}
    \item{\code{X1707}}{a numeric vector}
    \item{\code{X1708}}{a numeric vector}
    \item{\code{X1709}}{a numeric vector}
    \item{\code{X1710}}{a numeric vector}
    \item{\code{X1711}}{a numeric vector}
    \item{\code{X1712}}{a numeric vector}
    \item{\code{X1713}}{a numeric vector}
    \item{\code{X1714}}{a numeric vector}
    \item{\code{X1715}}{a numeric vector}
    \item{\code{X1716}}{a numeric vector}
    \item{\code{X1717}}{a numeric vector}
    \item{\code{X1718}}{a numeric vector}
    \item{\code{X1719}}{a numeric vector}
    \item{\code{X1720}}{a numeric vector}
    \item{\code{X1721}}{a numeric vector}
    \item{\code{X1722}}{a numeric vector}
    \item{\code{X1723}}{a numeric vector}
    \item{\code{X1724}}{a numeric vector}
    \item{\code{X1725}}{a numeric vector}
    \item{\code{X1726}}{a numeric vector}
    \item{\code{X1727}}{a numeric vector}
    \item{\code{X1728}}{a numeric vector}
    \item{\code{X1729}}{a numeric vector}
    \item{\code{X1730}}{a numeric vector}
    \item{\code{X1731}}{a numeric vector}
    \item{\code{X1732}}{a numeric vector}
    \item{\code{X1733}}{a numeric vector}
    \item{\code{X1734}}{a numeric vector}
    \item{\code{X1735}}{a numeric vector}
    \item{\code{X1736}}{a numeric vector}
    \item{\code{X1737}}{a numeric vector}
    \item{\code{X1738}}{a numeric vector}
    \item{\code{X1739}}{a numeric vector}
    \item{\code{X1740}}{a numeric vector}
    \item{\code{X1741}}{a numeric vector}
    \item{\code{X1742}}{a numeric vector}
    \item{\code{X1743}}{a numeric vector}
    \item{\code{X1744}}{a numeric vector}
    \item{\code{X1745}}{a numeric vector}
    \item{\code{X1746}}{a numeric vector}
    \item{\code{X1747}}{a numeric vector}
    \item{\code{X1748}}{a numeric vector}
    \item{\code{X1749}}{a numeric vector}
    \item{\code{X1750}}{a numeric vector}
    \item{\code{X1751}}{a numeric vector}
    \item{\code{X1752}}{a numeric vector}
    \item{\code{X1753}}{a numeric vector}
    \item{\code{X1754}}{a numeric vector}
    \item{\code{X1755}}{a numeric vector}
    \item{\code{X1756}}{a numeric vector}
    \item{\code{X1757}}{a numeric vector}
    \item{\code{X1758}}{a numeric vector}
    \item{\code{X1759}}{a numeric vector}
    \item{\code{X1760}}{a numeric vector}
    \item{\code{X1761}}{a numeric vector}
    \item{\code{X1762}}{a numeric vector}
    \item{\code{X1763}}{a numeric vector}
    \item{\code{X1764}}{a numeric vector}
    \item{\code{X1765}}{a numeric vector}
    \item{\code{X1766}}{a numeric vector}
    \item{\code{X1767}}{a numeric vector}
    \item{\code{X1768}}{a numeric vector}
    \item{\code{X1769}}{a numeric vector}
    \item{\code{X1770}}{a numeric vector}
    \item{\code{X1771}}{a numeric vector}
    \item{\code{X1772}}{a numeric vector}
    \item{\code{X1773}}{a numeric vector}
    \item{\code{X1774}}{a numeric vector}
    \item{\code{X1775}}{a numeric vector}
    \item{\code{X1776}}{a numeric vector}
    \item{\code{X1777}}{a numeric vector}
    \item{\code{X1778}}{a numeric vector}
    \item{\code{X1779}}{a numeric vector}
    \item{\code{X1780}}{a numeric vector}
    \item{\code{X1781}}{a numeric vector}
    \item{\code{X1782}}{a numeric vector}
    \item{\code{X1783}}{a numeric vector}
    \item{\code{X1784}}{a numeric vector}
    \item{\code{X1785}}{a numeric vector}
    \item{\code{X1786}}{a numeric vector}
    \item{\code{X1787}}{a numeric vector}
    \item{\code{X1788}}{a numeric vector}
    \item{\code{X1789}}{a numeric vector}
    \item{\code{X1790}}{a numeric vector}
    \item{\code{X1791}}{a numeric vector}
    \item{\code{X1792}}{a numeric vector}
    \item{\code{X1793}}{a numeric vector}
    \item{\code{X1794}}{a numeric vector}
    \item{\code{X1795}}{a numeric vector}
    \item{\code{X1796}}{a numeric vector}
    \item{\code{X1797}}{a numeric vector}
    \item{\code{X1798}}{a numeric vector}
    \item{\code{X1799}}{a numeric vector}
    \item{\code{X1800}}{a numeric vector}
    \item{\code{X1801}}{a numeric vector}
    \item{\code{X1802}}{a numeric vector}
    \item{\code{X1803}}{a numeric vector}
    \item{\code{X1804}}{a numeric vector}
    \item{\code{X1805}}{a numeric vector}
    \item{\code{X1806}}{a numeric vector}
    \item{\code{X1807}}{a numeric vector}
    \item{\code{X1808}}{a numeric vector}
    \item{\code{X1809}}{a numeric vector}
    \item{\code{X1810}}{a numeric vector}
    \item{\code{X1811}}{a numeric vector}
    \item{\code{X1812}}{a numeric vector}
    \item{\code{X1813}}{a numeric vector}
    \item{\code{X1814}}{a numeric vector}
    \item{\code{X1815}}{a numeric vector}
    \item{\code{X1816}}{a numeric vector}
    \item{\code{X1817}}{a numeric vector}
    \item{\code{X1818}}{a numeric vector}
    \item{\code{X1819}}{a numeric vector}
    \item{\code{X1820}}{a numeric vector}
    \item{\code{X1821}}{a numeric vector}
    \item{\code{X1822}}{a numeric vector}
    \item{\code{X1823}}{a numeric vector}
    \item{\code{X1824}}{a numeric vector}
    \item{\code{X1825}}{a numeric vector}
    \item{\code{X1826}}{a numeric vector}
    \item{\code{X1827}}{a numeric vector}
    \item{\code{X1828}}{a numeric vector}
    \item{\code{X1829}}{a numeric vector}
    \item{\code{X1830}}{a numeric vector}
    \item{\code{X1831}}{a numeric vector}
    \item{\code{X1832}}{a numeric vector}
    \item{\code{X1833}}{a numeric vector}
    \item{\code{X1834}}{a numeric vector}
    \item{\code{X1835}}{a numeric vector}
    \item{\code{X1836}}{a numeric vector}
    \item{\code{X1837}}{a numeric vector}
    \item{\code{X1838}}{a numeric vector}
    \item{\code{X1839}}{a numeric vector}
    \item{\code{X1840}}{a numeric vector}
    \item{\code{X1841}}{a numeric vector}
    \item{\code{X1842}}{a numeric vector}
    \item{\code{X1843}}{a numeric vector}
    \item{\code{X1844}}{a numeric vector}
    \item{\code{X1845}}{a numeric vector}
    \item{\code{X1846}}{a numeric vector}
    \item{\code{X1847}}{a numeric vector}
    \item{\code{X1848}}{a numeric vector}
    \item{\code{X1849}}{a numeric vector}
    \item{\code{X1850}}{a numeric vector}
    \item{\code{X1851}}{a numeric vector}
    \item{\code{X1852}}{a numeric vector}
    \item{\code{X1853}}{a numeric vector}
    \item{\code{X1854}}{a numeric vector}
    \item{\code{X1855}}{a numeric vector}
    \item{\code{X1856}}{a numeric vector}
    \item{\code{X1857}}{a numeric vector}
    \item{\code{X1858}}{a numeric vector}
    \item{\code{X1859}}{a numeric vector}
    \item{\code{X1860}}{a numeric vector}
    \item{\code{X1861}}{a numeric vector}
    \item{\code{X1862}}{a numeric vector}
    \item{\code{X1863}}{a numeric vector}
    \item{\code{X1864}}{a numeric vector}
    \item{\code{X1865}}{a numeric vector}
    \item{\code{X1866}}{a numeric vector}
    \item{\code{X1867}}{a numeric vector}
    \item{\code{X1868}}{a numeric vector}
    \item{\code{X1869}}{a numeric vector}
    \item{\code{X1870}}{a numeric vector}
    \item{\code{X1871}}{a numeric vector}
    \item{\code{X1872}}{a numeric vector}
    \item{\code{X1873}}{a numeric vector}
    \item{\code{X1874}}{a numeric vector}
    \item{\code{X1875}}{a numeric vector}
    \item{\code{X1876}}{a numeric vector}
    \item{\code{X1877}}{a numeric vector}
    \item{\code{X1878}}{a numeric vector}
    \item{\code{X1879}}{a numeric vector}
    \item{\code{X1880}}{a numeric vector}
    \item{\code{X1881}}{a numeric vector}
    \item{\code{X1882}}{a numeric vector}
    \item{\code{X1883}}{a numeric vector}
    \item{\code{X1884}}{a numeric vector}
    \item{\code{X1885}}{a numeric vector}
    \item{\code{X1886}}{a numeric vector}
    \item{\code{X1887}}{a numeric vector}
    \item{\code{X1888}}{a numeric vector}
    \item{\code{X1889}}{a numeric vector}
    \item{\code{X1890}}{a numeric vector}
    \item{\code{X1891}}{a numeric vector}
    \item{\code{X1892}}{a numeric vector}
    \item{\code{X1893}}{a numeric vector}
    \item{\code{X1894}}{a numeric vector}
    \item{\code{X1895}}{a numeric vector}
    \item{\code{X1896}}{a numeric vector}
    \item{\code{X1897}}{a numeric vector}
    \item{\code{X1898}}{a numeric vector}
    \item{\code{X1899}}{a numeric vector}
    \item{\code{X1900}}{a numeric vector}
    \item{\code{X1901}}{a numeric vector}
    \item{\code{X1902}}{a numeric vector}
    \item{\code{X1903}}{a numeric vector}
    \item{\code{X1904}}{a numeric vector}
    \item{\code{X1905}}{a numeric vector}
    \item{\code{X1906}}{a numeric vector}
    \item{\code{X1907}}{a numeric vector}
    \item{\code{X1908}}{a numeric vector}
    \item{\code{X1909}}{a numeric vector}
    \item{\code{X1910}}{a numeric vector}
    \item{\code{X1911}}{a numeric vector}
    \item{\code{X1912}}{a numeric vector}
    \item{\code{X1913}}{a numeric vector}
    \item{\code{X1914}}{a numeric vector}
    \item{\code{X1915}}{a numeric vector}
    \item{\code{X1916}}{a numeric vector}
    \item{\code{X1917}}{a numeric vector}
    \item{\code{X1918}}{a numeric vector}
    \item{\code{X1919}}{a numeric vector}
    \item{\code{X1920}}{a numeric vector}
    \item{\code{X1921}}{a numeric vector}
    \item{\code{X1922}}{a numeric vector}
    \item{\code{X1923}}{a numeric vector}
    \item{\code{X1924}}{a numeric vector}
    \item{\code{X1925}}{a numeric vector}
    \item{\code{X1926}}{a numeric vector}
    \item{\code{X1927}}{a numeric vector}
    \item{\code{X1928}}{a numeric vector}
    \item{\code{X1929}}{a numeric vector}
    \item{\code{X1930}}{a numeric vector}
    \item{\code{X1931}}{a numeric vector}
    \item{\code{X1932}}{a numeric vector}
    \item{\code{X1933}}{a numeric vector}
    \item{\code{X1934}}{a numeric vector}
    \item{\code{X1935}}{a numeric vector}
    \item{\code{X1936}}{a numeric vector}
    \item{\code{X1937}}{a numeric vector}
    \item{\code{X1938}}{a numeric vector}
    \item{\code{X1939}}{a numeric vector}
    \item{\code{X1940}}{a numeric vector}
    \item{\code{X1941}}{a numeric vector}
    \item{\code{X1942}}{a numeric vector}
    \item{\code{X1943}}{a numeric vector}
    \item{\code{X1944}}{a numeric vector}
    \item{\code{X1945}}{a numeric vector}
    \item{\code{X1946}}{a numeric vector}
    \item{\code{X1947}}{a numeric vector}
    \item{\code{X1948}}{a numeric vector}
    \item{\code{X1949}}{a numeric vector}
    \item{\code{X1950}}{a numeric vector}
    \item{\code{X1951}}{a numeric vector}
    \item{\code{X1952}}{a numeric vector}
    \item{\code{X1953}}{a numeric vector}
    \item{\code{X1954}}{a numeric vector}
    \item{\code{X1955}}{a numeric vector}
    \item{\code{X1956}}{a numeric vector}
    \item{\code{X1957}}{a numeric vector}
    \item{\code{X1958}}{a numeric vector}
    \item{\code{X1959}}{a numeric vector}
    \item{\code{X1960}}{a numeric vector}
    \item{\code{X1961}}{a numeric vector}
    \item{\code{X1962}}{a numeric vector}
    \item{\code{X1963}}{a numeric vector}
    \item{\code{X1964}}{a numeric vector}
    \item{\code{X1965}}{a numeric vector}
    \item{\code{X1966}}{a numeric vector}
    \item{\code{X1967}}{a numeric vector}
    \item{\code{X1968}}{a numeric vector}
    \item{\code{X1969}}{a numeric vector}
    \item{\code{X1970}}{a numeric vector}
    \item{\code{X1971}}{a numeric vector}
    \item{\code{X1972}}{a numeric vector}
    \item{\code{X1973}}{a numeric vector}
    \item{\code{X1974}}{a numeric vector}
    \item{\code{X1975}}{a numeric vector}
    \item{\code{X1976}}{a numeric vector}
    \item{\code{X1977}}{a numeric vector}
    \item{\code{X1978}}{a numeric vector}
    \item{\code{X1979}}{a numeric vector}
    \item{\code{X1980}}{a numeric vector}
    \item{\code{X1981}}{a numeric vector}
    \item{\code{X1982}}{a numeric vector}
    \item{\code{X1983}}{a numeric vector}
    \item{\code{X1984}}{a numeric vector}
    \item{\code{X1985}}{a numeric vector}
    \item{\code{X1986}}{a numeric vector}
    \item{\code{X1987}}{a numeric vector}
    \item{\code{X1988}}{a numeric vector}
    \item{\code{X1989}}{a numeric vector}
    \item{\code{X1990}}{a numeric vector}
    \item{\code{X1991}}{a numeric vector}
    \item{\code{X1992}}{a numeric vector}
    \item{\code{X1993}}{a numeric vector}
    \item{\code{X1994}}{a numeric vector}
    \item{\code{X1995}}{a numeric vector}
    \item{\code{X1996}}{a numeric vector}
    \item{\code{X1997}}{a numeric vector}
    \item{\code{X1998}}{a numeric vector}
    \item{\code{X1999}}{a numeric vector}
    \item{\code{X2000}}{a numeric vector}
    \item{\code{X2001}}{a numeric vector}
    \item{\code{X2002}}{a numeric vector}
    \item{\code{X2003}}{a numeric vector}
    \item{\code{X2004}}{a numeric vector}
    \item{\code{X2005}}{a numeric vector}
    \item{\code{X2006}}{a numeric vector}
    \item{\code{X2007}}{a numeric vector}
    \item{\code{X2008}}{a numeric vector}
    \item{\code{X2009}}{a numeric vector}
    \item{\code{X2010}}{a numeric vector}
    \item{\code{X2011}}{a numeric vector}
    \item{\code{X2012}}{a numeric vector}
    \item{\code{X2013}}{a numeric vector}
    \item{\code{X2014}}{a numeric vector}
    \item{\code{X2015}}{a numeric vector}
    \item{\code{X2016}}{a numeric vector}
    \item{\code{X2017}}{a numeric vector}
    \item{\code{X2018}}{a numeric vector}
    \item{\code{X2019}}{a numeric vector}
    \item{\code{X2020}}{a numeric vector}
    \item{\code{X2021}}{a numeric vector}
    \item{\code{X2022}}{a numeric vector}
    \item{\code{X2023}}{a numeric vector}
    \item{\code{X2024}}{a numeric vector}
    \item{\code{X2025}}{a numeric vector}
    \item{\code{X2026}}{a numeric vector}
    \item{\code{X2027}}{a numeric vector}
    \item{\code{X2028}}{a numeric vector}
    \item{\code{X2029}}{a numeric vector}
    \item{\code{X2030}}{a numeric vector}
    \item{\code{X2031}}{a numeric vector}
    \item{\code{X2032}}{a numeric vector}
    \item{\code{X2033}}{a numeric vector}
    \item{\code{X2034}}{a numeric vector}
    \item{\code{X2035}}{a numeric vector}
    \item{\code{X2036}}{a numeric vector}
    \item{\code{X2037}}{a numeric vector}
    \item{\code{X2038}}{a numeric vector}
    \item{\code{X2039}}{a numeric vector}
    \item{\code{X2040}}{a numeric vector}
    \item{\code{X2041}}{a numeric vector}
    \item{\code{X2042}}{a numeric vector}
    \item{\code{X2043}}{a numeric vector}
    \item{\code{X2044}}{a numeric vector}
    \item{\code{X2045}}{a numeric vector}
    \item{\code{X2046}}{a numeric vector}
    \item{\code{X2047}}{a numeric vector}
    \item{\code{X2048}}{a numeric vector}
    \item{\code{X2049}}{a numeric vector}
    \item{\code{X2050}}{a numeric vector}
    \item{\code{X2051}}{a numeric vector}
    \item{\code{X2052}}{a numeric vector}
    \item{\code{X2053}}{a numeric vector}
    \item{\code{X2054}}{a numeric vector}
    \item{\code{X2055}}{a numeric vector}
    \item{\code{X2056}}{a numeric vector}
    \item{\code{X2057}}{a numeric vector}
    \item{\code{X2058}}{a numeric vector}
    \item{\code{X2059}}{a numeric vector}
    \item{\code{X2060}}{a numeric vector}
    \item{\code{X2061}}{a numeric vector}
    \item{\code{X2062}}{a numeric vector}
    \item{\code{X2063}}{a numeric vector}
    \item{\code{X2064}}{a numeric vector}
    \item{\code{X2065}}{a numeric vector}
    \item{\code{X2066}}{a numeric vector}
    \item{\code{X2067}}{a numeric vector}
    \item{\code{X2068}}{a numeric vector}
    \item{\code{X2069}}{a numeric vector}
    \item{\code{X2070}}{a numeric vector}
    \item{\code{X2071}}{a numeric vector}
    \item{\code{X2072}}{a numeric vector}
    \item{\code{X2073}}{a numeric vector}
    \item{\code{X2074}}{a numeric vector}
    \item{\code{X2075}}{a numeric vector}
    \item{\code{X2076}}{a numeric vector}
    \item{\code{X2077}}{a numeric vector}
    \item{\code{X2078}}{a numeric vector}
    \item{\code{X2079}}{a numeric vector}
    \item{\code{X2080}}{a numeric vector}
    \item{\code{X2081}}{a numeric vector}
    \item{\code{X2082}}{a numeric vector}
    \item{\code{X2083}}{a numeric vector}
    \item{\code{X2084}}{a numeric vector}
    \item{\code{X2085}}{a numeric vector}
    \item{\code{X2086}}{a numeric vector}
    \item{\code{X2087}}{a numeric vector}
    \item{\code{X2088}}{a numeric vector}
    \item{\code{X2089}}{a numeric vector}
    \item{\code{X2090}}{a numeric vector}
    \item{\code{X2091}}{a numeric vector}
    \item{\code{X2092}}{a numeric vector}
    \item{\code{X2093}}{a numeric vector}
    \item{\code{X2094}}{a numeric vector}
    \item{\code{X2095}}{a numeric vector}
    \item{\code{X2096}}{a numeric vector}
    \item{\code{X2097}}{a numeric vector}
    \item{\code{X2098}}{a numeric vector}
    \item{\code{X2099}}{a numeric vector}
    \item{\code{X2100}}{a numeric vector}
    \item{\code{X2101}}{a numeric vector}
    \item{\code{X2102}}{a numeric vector}
    \item{\code{X2103}}{a numeric vector}
    \item{\code{X2104}}{a numeric vector}
    \item{\code{X2105}}{a numeric vector}
    \item{\code{X2106}}{a numeric vector}
    \item{\code{X2107}}{a numeric vector}
    \item{\code{X2108}}{a numeric vector}
    \item{\code{X2109}}{a numeric vector}
    \item{\code{X2110}}{a numeric vector}
    \item{\code{X2111}}{a numeric vector}
    \item{\code{X2112}}{a numeric vector}
    \item{\code{X2113}}{a numeric vector}
    \item{\code{X2114}}{a numeric vector}
    \item{\code{X2115}}{a numeric vector}
    \item{\code{X2116}}{a numeric vector}
    \item{\code{X2117}}{a numeric vector}
    \item{\code{X2118}}{a numeric vector}
    \item{\code{X2119}}{a numeric vector}
    \item{\code{X2120}}{a numeric vector}
    \item{\code{X2121}}{a numeric vector}
    \item{\code{X2122}}{a numeric vector}
    \item{\code{X2123}}{a numeric vector}
    \item{\code{X2124}}{a numeric vector}
    \item{\code{X2125}}{a numeric vector}
    \item{\code{X2126}}{a numeric vector}
    \item{\code{X2127}}{a numeric vector}
    \item{\code{X2128}}{a numeric vector}
    \item{\code{X2129}}{a numeric vector}
    \item{\code{X2130}}{a numeric vector}
    \item{\code{X2131}}{a numeric vector}
    \item{\code{X2132}}{a numeric vector}
    \item{\code{X2133}}{a numeric vector}
    \item{\code{X2134}}{a numeric vector}
    \item{\code{X2135}}{a numeric vector}
    \item{\code{X2136}}{a numeric vector}
    \item{\code{X2137}}{a numeric vector}
    \item{\code{X2138}}{a numeric vector}
    \item{\code{X2139}}{a numeric vector}
    \item{\code{X2140}}{a numeric vector}
    \item{\code{X2141}}{a numeric vector}
    \item{\code{X2142}}{a numeric vector}
    \item{\code{X2143}}{a numeric vector}
    \item{\code{X2144}}{a numeric vector}
    \item{\code{X2145}}{a numeric vector}
    \item{\code{X2146}}{a numeric vector}
    \item{\code{X2147}}{a numeric vector}
    \item{\code{X2148}}{a numeric vector}
    \item{\code{X2149}}{a numeric vector}
    \item{\code{X2150}}{a numeric vector}
    \item{\code{X2151}}{a numeric vector}
    \item{\code{X2152}}{a numeric vector}
    \item{\code{X2153}}{a numeric vector}
    \item{\code{X2154}}{a numeric vector}
    \item{\code{X2155}}{a numeric vector}
    \item{\code{X2156}}{a numeric vector}
    \item{\code{X2157}}{a numeric vector}
    \item{\code{X2158}}{a numeric vector}
    \item{\code{X2159}}{a numeric vector}
    \item{\code{X2160}}{a numeric vector}
    \item{\code{X2161}}{a numeric vector}
    \item{\code{X2162}}{a numeric vector}
    \item{\code{X2163}}{a numeric vector}
    \item{\code{X2164}}{a numeric vector}
    \item{\code{X2165}}{a numeric vector}
    \item{\code{X2166}}{a numeric vector}
    \item{\code{X2167}}{a numeric vector}
    \item{\code{X2168}}{a numeric vector}
    \item{\code{X2169}}{a numeric vector}
    \item{\code{X2170}}{a numeric vector}
    \item{\code{X2171}}{a numeric vector}
    \item{\code{X2172}}{a numeric vector}
    \item{\code{X2173}}{a numeric vector}
    \item{\code{X2174}}{a numeric vector}
    \item{\code{X2175}}{a numeric vector}
    \item{\code{X2176}}{a numeric vector}
    \item{\code{X2177}}{a numeric vector}
    \item{\code{X2178}}{a numeric vector}
    \item{\code{X2179}}{a numeric vector}
    \item{\code{X2180}}{a numeric vector}
    \item{\code{X2181}}{a numeric vector}
    \item{\code{X2182}}{a numeric vector}
    \item{\code{X2183}}{a numeric vector}
    \item{\code{X2184}}{a numeric vector}
    \item{\code{X2185}}{a numeric vector}
    \item{\code{X2186}}{a numeric vector}
    \item{\code{X2187}}{a numeric vector}
    \item{\code{X2188}}{a numeric vector}
    \item{\code{X2189}}{a numeric vector}
    \item{\code{X2190}}{a numeric vector}
    \item{\code{X2191}}{a numeric vector}
    \item{\code{X2192}}{a numeric vector}
    \item{\code{X2193}}{a numeric vector}
    \item{\code{X2194}}{a numeric vector}
    \item{\code{X2195}}{a numeric vector}
    \item{\code{X2196}}{a numeric vector}
    \item{\code{X2197}}{a numeric vector}
    \item{\code{X2198}}{a numeric vector}
    \item{\code{X2199}}{a numeric vector}
    \item{\code{X2200}}{a numeric vector}
    \item{\code{X2201}}{a numeric vector}
    \item{\code{X2202}}{a numeric vector}
    \item{\code{X2203}}{a numeric vector}
    \item{\code{X2204}}{a numeric vector}
    \item{\code{X2205}}{a numeric vector}
    \item{\code{X2206}}{a numeric vector}
    \item{\code{X2207}}{a numeric vector}
    \item{\code{X2208}}{a numeric vector}
    \item{\code{X2209}}{a numeric vector}
    \item{\code{X2210}}{a numeric vector}
    \item{\code{X2211}}{a numeric vector}
    \item{\code{X2212}}{a numeric vector}
    \item{\code{X2213}}{a numeric vector}
    \item{\code{X2214}}{a numeric vector}
    \item{\code{X2215}}{a numeric vector}
    \item{\code{X2216}}{a numeric vector}
    \item{\code{X2217}}{a numeric vector}
    \item{\code{X2218}}{a numeric vector}
    \item{\code{X2219}}{a numeric vector}
    \item{\code{X2220}}{a numeric vector}
    \item{\code{X2221}}{a numeric vector}
    \item{\code{X2222}}{a numeric vector}
    \item{\code{X2223}}{a numeric vector}
    \item{\code{X2224}}{a numeric vector}
    \item{\code{X2225}}{a numeric vector}
    \item{\code{X2226}}{a numeric vector}
    \item{\code{X2227}}{a numeric vector}
    \item{\code{X2228}}{a numeric vector}
    \item{\code{X2229}}{a numeric vector}
    \item{\code{X2230}}{a numeric vector}
    \item{\code{X2231}}{a numeric vector}
    \item{\code{X2232}}{a numeric vector}
    \item{\code{X2233}}{a numeric vector}
    \item{\code{X2234}}{a numeric vector}
    \item{\code{X2235}}{a numeric vector}
    \item{\code{X2236}}{a numeric vector}
    \item{\code{X2237}}{a numeric vector}
    \item{\code{X2238}}{a numeric vector}
    \item{\code{X2239}}{a numeric vector}
    \item{\code{X2240}}{a numeric vector}
    \item{\code{X2241}}{a numeric vector}
    \item{\code{X2242}}{a numeric vector}
    \item{\code{X2243}}{a numeric vector}
    \item{\code{X2244}}{a numeric vector}
    \item{\code{X2245}}{a numeric vector}
    \item{\code{X2246}}{a numeric vector}
    \item{\code{X2247}}{a numeric vector}
    \item{\code{X2248}}{a numeric vector}
    \item{\code{X2249}}{a numeric vector}
    \item{\code{X2250}}{a numeric vector}
    \item{\code{X2251}}{a numeric vector}
    \item{\code{X2252}}{a numeric vector}
    \item{\code{X2253}}{a numeric vector}
    \item{\code{X2254}}{a numeric vector}
    \item{\code{X2255}}{a numeric vector}
    \item{\code{X2256}}{a numeric vector}
    \item{\code{X2257}}{a numeric vector}
    \item{\code{X2258}}{a numeric vector}
    \item{\code{X2259}}{a numeric vector}
    \item{\code{X2260}}{a numeric vector}
    \item{\code{X2261}}{a numeric vector}
    \item{\code{X2262}}{a numeric vector}
    \item{\code{X2263}}{a numeric vector}
    \item{\code{X2264}}{a numeric vector}
    \item{\code{X2265}}{a numeric vector}
    \item{\code{X2266}}{a numeric vector}
    \item{\code{X2267}}{a numeric vector}
    \item{\code{X2268}}{a numeric vector}
    \item{\code{X2269}}{a numeric vector}
    \item{\code{X2270}}{a numeric vector}
    \item{\code{X2271}}{a numeric vector}
    \item{\code{X2272}}{a numeric vector}
    \item{\code{X2273}}{a numeric vector}
    \item{\code{X2274}}{a numeric vector}
    \item{\code{X2275}}{a numeric vector}
    \item{\code{X2276}}{a numeric vector}
    \item{\code{X2277}}{a numeric vector}
    \item{\code{X2278}}{a numeric vector}
    \item{\code{X2279}}{a numeric vector}
    \item{\code{X2280}}{a numeric vector}
    \item{\code{X2281}}{a numeric vector}
    \item{\code{X2282}}{a numeric vector}
    \item{\code{X2283}}{a numeric vector}
    \item{\code{X2284}}{a numeric vector}
    \item{\code{X2285}}{a numeric vector}
    \item{\code{X2286}}{a numeric vector}
    \item{\code{X2287}}{a numeric vector}
    \item{\code{X2288}}{a numeric vector}
    \item{\code{X2289}}{a numeric vector}
    \item{\code{X2290}}{a numeric vector}
    \item{\code{X2291}}{a numeric vector}
    \item{\code{X2292}}{a numeric vector}
    \item{\code{X2293}}{a numeric vector}
    \item{\code{X2294}}{a numeric vector}
    \item{\code{X2295}}{a numeric vector}
    \item{\code{X2296}}{a numeric vector}
    \item{\code{X2297}}{a numeric vector}
    \item{\code{X2298}}{a numeric vector}
    \item{\code{X2299}}{a numeric vector}
    \item{\code{X2300}}{a numeric vector}
    \item{\code{X2301}}{a numeric vector}
    \item{\code{X2302}}{a numeric vector}
    \item{\code{X2303}}{a numeric vector}
    \item{\code{X2304}}{a numeric vector}
    \item{\code{X2305}}{a numeric vector}
    \item{\code{X2306}}{a numeric vector}
    \item{\code{X2307}}{a numeric vector}
    \item{\code{X2308}}{a numeric vector}
    \item{\code{X2309}}{a numeric vector}
    \item{\code{X2310}}{a numeric vector}
    \item{\code{X2311}}{a numeric vector}
    \item{\code{X2312}}{a numeric vector}
    \item{\code{X2313}}{a numeric vector}
    \item{\code{X2314}}{a numeric vector}
    \item{\code{X2315}}{a numeric vector}
    \item{\code{X2316}}{a numeric vector}
    \item{\code{X2317}}{a numeric vector}
    \item{\code{X2318}}{a numeric vector}
    \item{\code{X2319}}{a numeric vector}
    \item{\code{X2320}}{a numeric vector}
    \item{\code{X2321}}{a numeric vector}
    \item{\code{X2322}}{a numeric vector}
    \item{\code{X2323}}{a numeric vector}
    \item{\code{X2324}}{a numeric vector}
    \item{\code{X2325}}{a numeric vector}
    \item{\code{X2326}}{a numeric vector}
    \item{\code{X2327}}{a numeric vector}
    \item{\code{X2328}}{a numeric vector}
    \item{\code{X2329}}{a numeric vector}
    \item{\code{X2330}}{a numeric vector}
    \item{\code{X2331}}{a numeric vector}
    \item{\code{X2332}}{a numeric vector}
    \item{\code{X2333}}{a numeric vector}
    \item{\code{X2334}}{a numeric vector}
    \item{\code{X2335}}{a numeric vector}
    \item{\code{X2336}}{a numeric vector}
    \item{\code{X2337}}{a numeric vector}
    \item{\code{X2338}}{a numeric vector}
    \item{\code{X2339}}{a numeric vector}
    \item{\code{X2340}}{a numeric vector}
    \item{\code{X2341}}{a numeric vector}
    \item{\code{X2342}}{a numeric vector}
    \item{\code{X2343}}{a numeric vector}
    \item{\code{X2344}}{a numeric vector}
    \item{\code{X2345}}{a numeric vector}
    \item{\code{X2346}}{a numeric vector}
    \item{\code{X2347}}{a numeric vector}
    \item{\code{X2348}}{a numeric vector}
    \item{\code{X2349}}{a numeric vector}
    \item{\code{X2350}}{a numeric vector}
    \item{\code{X2351}}{a numeric vector}
    \item{\code{X2352}}{a numeric vector}
    \item{\code{X2353}}{a numeric vector}
    \item{\code{X2354}}{a numeric vector}
    \item{\code{X2355}}{a numeric vector}
    \item{\code{X2356}}{a numeric vector}
    \item{\code{X2357}}{a numeric vector}
    \item{\code{X2358}}{a numeric vector}
    \item{\code{X2359}}{a numeric vector}
    \item{\code{X2360}}{a numeric vector}
    \item{\code{X2361}}{a numeric vector}
    \item{\code{X2362}}{a numeric vector}
    \item{\code{X2363}}{a numeric vector}
    \item{\code{X2364}}{a numeric vector}
    \item{\code{X2365}}{a numeric vector}
    \item{\code{X2366}}{a numeric vector}
    \item{\code{X2367}}{a numeric vector}
    \item{\code{X2368}}{a numeric vector}
    \item{\code{X2369}}{a numeric vector}
    \item{\code{X2370}}{a numeric vector}
    \item{\code{X2371}}{a numeric vector}
    \item{\code{X2372}}{a numeric vector}
    \item{\code{X2373}}{a numeric vector}
    \item{\code{X2374}}{a numeric vector}
    \item{\code{X2375}}{a numeric vector}
    \item{\code{X2376}}{a numeric vector}
    \item{\code{X2377}}{a numeric vector}
    \item{\code{X2378}}{a numeric vector}
    \item{\code{X2379}}{a numeric vector}
    \item{\code{X2380}}{a numeric vector}
    \item{\code{X2381}}{a numeric vector}
    \item{\code{X2382}}{a numeric vector}
    \item{\code{X2383}}{a numeric vector}
    \item{\code{X2384}}{a numeric vector}
    \item{\code{X2385}}{a numeric vector}
    \item{\code{X2386}}{a numeric vector}
    \item{\code{X2387}}{a numeric vector}
    \item{\code{X2388}}{a numeric vector}
    \item{\code{X2389}}{a numeric vector}
    \item{\code{X2390}}{a numeric vector}
    \item{\code{X2391}}{a numeric vector}
    \item{\code{X2392}}{a numeric vector}
    \item{\code{X2393}}{a numeric vector}
    \item{\code{X2394}}{a numeric vector}
    \item{\code{X2395}}{a numeric vector}
    \item{\code{X2396}}{a numeric vector}
    \item{\code{X2397}}{a numeric vector}
    \item{\code{X2398}}{a numeric vector}
    \item{\code{X2399}}{a numeric vector}
    \item{\code{X2400}}{a numeric vector}
    \item{\code{X2401}}{a numeric vector}
    \item{\code{X2402}}{a numeric vector}
    \item{\code{X2403}}{a numeric vector}
    \item{\code{X2404}}{a numeric vector}
    \item{\code{X2405}}{a numeric vector}
    \item{\code{X2406}}{a numeric vector}
    \item{\code{X2407}}{a numeric vector}
    \item{\code{X2408}}{a numeric vector}
    \item{\code{X2409}}{a numeric vector}
    \item{\code{X2410}}{a numeric vector}
    \item{\code{X2411}}{a numeric vector}
    \item{\code{X2412}}{a numeric vector}
    \item{\code{X2413}}{a numeric vector}
    \item{\code{X2414}}{a numeric vector}
    \item{\code{X2415}}{a numeric vector}
    \item{\code{X2416}}{a numeric vector}
    \item{\code{X2417}}{a numeric vector}
    \item{\code{X2418}}{a numeric vector}
    \item{\code{X2419}}{a numeric vector}
    \item{\code{X2420}}{a numeric vector}
    \item{\code{X2421}}{a numeric vector}
    \item{\code{X2422}}{a numeric vector}
    \item{\code{X2423}}{a numeric vector}
    \item{\code{X2424}}{a numeric vector}
    \item{\code{X2425}}{a numeric vector}
    \item{\code{X2426}}{a numeric vector}
    \item{\code{X2427}}{a numeric vector}
    \item{\code{X2428}}{a numeric vector}
    \item{\code{X2429}}{a numeric vector}
    \item{\code{X2430}}{a numeric vector}
    \item{\code{X2431}}{a numeric vector}
    \item{\code{X2432}}{a numeric vector}
    \item{\code{X2433}}{a numeric vector}
    \item{\code{X2434}}{a numeric vector}
    \item{\code{X2435}}{a numeric vector}
    \item{\code{X2436}}{a numeric vector}
    \item{\code{X2437}}{a numeric vector}
    \item{\code{X2438}}{a numeric vector}
    \item{\code{X2439}}{a numeric vector}
    \item{\code{X2440}}{a numeric vector}
    \item{\code{X2441}}{a numeric vector}
    \item{\code{X2442}}{a numeric vector}
    \item{\code{X2443}}{a numeric vector}
    \item{\code{X2444}}{a numeric vector}
    \item{\code{X2445}}{a numeric vector}
    \item{\code{X2446}}{a numeric vector}
    \item{\code{X2447}}{a numeric vector}
    \item{\code{X2448}}{a numeric vector}
    \item{\code{X2449}}{a numeric vector}
    \item{\code{X2450}}{a numeric vector}
    \item{\code{X2451}}{a numeric vector}
    \item{\code{X2452}}{a numeric vector}
    \item{\code{X2453}}{a numeric vector}
    \item{\code{X2454}}{a numeric vector}
    \item{\code{X2455}}{a numeric vector}
    \item{\code{X2456}}{a numeric vector}
    \item{\code{X2457}}{a numeric vector}
    \item{\code{X2458}}{a numeric vector}
    \item{\code{X2459}}{a numeric vector}
    \item{\code{X2460}}{a numeric vector}
    \item{\code{X2461}}{a numeric vector}
    \item{\code{X2462}}{a numeric vector}
    \item{\code{X2463}}{a numeric vector}
    \item{\code{X2464}}{a numeric vector}
    \item{\code{X2465}}{a numeric vector}
    \item{\code{X2466}}{a numeric vector}
    \item{\code{X2467}}{a numeric vector}
    \item{\code{X2468}}{a numeric vector}
    \item{\code{X2469}}{a numeric vector}
    \item{\code{X2470}}{a numeric vector}
    \item{\code{X2471}}{a numeric vector}
    \item{\code{X2472}}{a numeric vector}
    \item{\code{X2473}}{a numeric vector}
    \item{\code{X2474}}{a numeric vector}
    \item{\code{X2475}}{a numeric vector}
    \item{\code{X2476}}{a numeric vector}
    \item{\code{X2477}}{a numeric vector}
    \item{\code{X2478}}{a numeric vector}
    \item{\code{X2479}}{a numeric vector}
    \item{\code{X2480}}{a numeric vector}
    \item{\code{X2481}}{a numeric vector}
    \item{\code{X2482}}{a numeric vector}
    \item{\code{X2483}}{a numeric vector}
    \item{\code{X2484}}{a numeric vector}
    \item{\code{X2485}}{a numeric vector}
    \item{\code{X2486}}{a numeric vector}
    \item{\code{X2487}}{a numeric vector}
    \item{\code{X2488}}{a numeric vector}
    \item{\code{X2489}}{a numeric vector}
    \item{\code{X2490}}{a numeric vector}
    \item{\code{X2491}}{a numeric vector}
    \item{\code{X2492}}{a numeric vector}
    \item{\code{X2493}}{a numeric vector}
    \item{\code{X2494}}{a numeric vector}
    \item{\code{X2495}}{a numeric vector}
    \item{\code{X2496}}{a numeric vector}
    \item{\code{X2497}}{a numeric vector}
    \item{\code{X2498}}{a numeric vector}
    \item{\code{X2499}}{a numeric vector}
    \item{\code{X2500}}{a numeric vector}
  }
}
\details{
Severe atmosphereic absorption present in some bansd. 
}
\source{
Hans D Roelofsen, KWR Watercycle Research Institute
}
\references{
Roelofsen etal, 2013, in preperation
}
\examples{
data(SampleHySpecRefl)
plot(as.numeric(substring(colnames(SampleHySpecRefl), 2)), apply(SampleHySpecRefl, 2, mean), ylim= c(0, 1), type= "l", xlab= "wavelength [nm]", ylab= "mean reflection")
}
\keyword{datasets}
