\name{SampleWavelengths}
\alias{SampleWavelengths}
\docType{data}
\title{
Wavelengths of 2151 spectral bands
}
\description{
Numeric vector with the wavelength of each of 2151 spectral bands that are standard output of an ASD FieldSpec PRO FR Field Spectrometer. 
}
\usage{data(SampleWavelengths)}
\format{
  The format is:
 num [1:2151] 350 351 352 353 354 355 356 357 358 359 ...
}
\details{
Unit= nm
}
\source{
Hans D Roelofsen, KWR Watercycle Research Institute
}
\references{
Roelofsen etal, 2013, in preperation
}
\examples{
data(SampleWavelengths)
summary(SampleWavelengths)
}
\keyword{datasets}
