% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdR-package.R
\docType{package}
\name{hdR-package}
\alias{hdR}
\alias{hdR-package}
\title{hdR: download data from Human Development Report Office (HDRO)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brunomioto/hdR}
  \item Report bugs at \url{https://github.com/brunomioto/hdR/issues}
}

}
\author{
\strong{Maintainer}: Bruno Mioto \email{brunomioto97@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
